% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addfvtomap.R
\name{addfvtomap}
\alias{addfvtomap}
\title{addfvtomap}
\usage{
addfvtomap(fvobj, mapxlim, mapylim)
}
\arguments{
\item{fvobj}{Is the fv function to plot}

\item{mapxlim}{is a vector of length 2 giving the x-coordinates of the region to use to plot \code{fvobj} in the units of the current figure.}

\item{mapylim}{Same as mapxlim except for the y-coordinates.}
}
\value{
Nothing. Adds a plot of the default function value in \code{fvobj} to the current figure within the region given by mapxlim and mapylim.
}
\description{
A function for plotting fv objects onto large spatial map figures. This allows localised summary functions to be plotted on the spatial region that they describe.
}
