% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varofobsarea.all.R
\name{allsae}
\alias{allsae}
\title{Variance Estimates for Observed Area}
\usage{
allsae(xi, obswin, corrrad, corrstepheight, erosionrad, n11, n21, n12, n22)
}
\arguments{
\item{xi}{An observation of the RACS of interest in 1, 0, or NA valued pixels.
Pixels must be square. Must be a spatstat im object.}

\item{obswin}{Observation window}

\item{corrrad}{Radius of the step function in the correlation (in the same units as xi)}

\item{corrstepheight}{Height of the step in the correlation}

\item{n11}{The number of samples that were truly class 1 and fallibly classified as class 1}

\item{n21}{The number of samples that were truly class 2 and fallibly classified as class 1}

\item{n12}{The number of samples that were truly class 1 and fallibly classified as class 2}

\item{n22}{The number of samples that were truly class 2 and fallibly classified as class 2}
}
\description{
Estimates the variance of the area of a cover type observed in a thematic map created using a fallible classifier from remote sensing.
}
\examples{
xi <- heather$coarse
obswin <- Frame(heather$coarse)
corrrad <- 2
corrstepheight <- 0.95
n11 <- 90
n21 <- 10
n22 <- 95
n12 <- 5
estimates <- allsae(xi, obswin, corrrad, corrstepheight, erosionrad = 5, n11, n21, n12, n22)
\dontrun{
plot.new()
plot.window(ylim = c(min(estimates$areahat - 2*sqrt(estimates$varhat)),
                     max(estimates$areahat + 2*sqrt(estimates$varhat))),
           xlim = c(0, length(estimates$areahat) + 1))
axis(1, at = 1:length(estimates$areahat), labels = row.names(estimates))
axis(2)
arrows(1:length(estimates$areahat),
  y0 = estimates$areahat - 2*sqrt(estimates$varhat),
  y1 = estimates$areahat + 2*sqrt(estimates$varhat),
  length = 0.05, angle = 90, code =3)
}
}
\author{
{Kassel Hingee}
}
