% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asondiskrasters.R
\name{as.rasterlayer.im}
\alias{as.rasterlayer.im}
\alias{offram}
\alias{imtoondiskras}
\alias{imstoondiskras}
\alias{rasterstoims}
\title{Functions for converting lists of im Ojects to on-disk rasterLayer objects}
\usage{
as.rasterlayer.im(im, layername = NULL)

offram(rasterobj, filenamesave = NULL)

imtoondiskras(im, layername = NULL, filenamesave = NULL)

imstoondiskras(alist, basefilename = NULL, recursive = FALSE)

rasterstoims(alist, recursive = FALSE)
}
\arguments{
\item{im}{A spatstat image object}

\item{layername}{Optional layer name to give to the created rasterLayer}

\item{rasterobj}{A Raster* object}

\item{filenamesave}{The filename to save \code{rasterobj} to.}

\item{alist}{A list containing some \code{im} objects.}

\item{basefilename}{The filename that is used as root of all filenames.}

\item{recursive}{If TRUE sublists of \code{alist} will also be modified.}
}
\description{
Functions for saving spatstat's \code{im} objects as \code{raster} objects out of RAM. Useful when computer has limited RAM and a fast solid state disk to store the images off RAM.
\code{as.rasterlayer.im} Converts a \code{spatstat} \code{im} into a \code{Raster} \code{rasterLayer}
}
\details{
These functions are useful when experiments generate a large collection of \code{im} objects.
The package \code{raster} has many useful tools for working with these images outside RAM.
I think it is wise to garbage collect \code{gc()} just after overwriting the im objects.
}
\section{Functions}{
\itemize{
\item \code{offram}: Create a rasterLayer (or other Raster* object) that is saved on disk out of another raster object.
If \code{filenamesave} is \code{NULL} then a temporary location is used.

\item \code{imtoondiskras}: Create a rasterLayer (or other Raster* object) that is saved on disk out of a \code{im} object.

\item \code{imstoondiskras}: Replaces all \code{im} elements in a list with on-disk \code{rasterLayer} objects.
Raster data is saved in \code{basefilename_[listentryname]}

\item \code{rasterstoims}: Replaces all code{rasterLayer} objects in \code{alist} with spatstat \code{im} objects. Uses MAPTOOLS
}}

\examples{
xi <- heather$coarse
cvchats <- balancedracscovariances(xi, Frame(xi), modifications = "all")
cvchatsondisk <- imstoondiskras(cvchats)
cvchaatsbackasim <- rasterstoims(cvchatsondisk)
}
