% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancedracscovariance.R
\name{balancedracscovariances}
\alias{balancedracscovariances}
\alias{balancedracscovariance.cvchat}
\alias{balancedracscovariances.cvchat}
\alias{cvchats.convolves}
\title{Balanced spatial covariance estimation, also known as `two-point probability', estimator for stationary RACS}
\usage{
balancedracscovariances(xi, obswin = NULL, setcov_boundarythresh = NULL,
  modifications = NULL)

balancedracscovariance.cvchat(cvchat, cpp1 = NULL, phat = NULL,
  modification = NULL)

balancedracscovariances.cvchat(cvchat, cpp1 = NULL, phat = NULL,
  modifications = NULL)

cvchats.convolves(xixi, winwin, xiwin = NULL, phat = NULL,
  modifications = NULL)
}
\arguments{
\item{xi}{An observation of the RACS of interest. It can be in \pkg{spatstat}'s \code{owin} or \code{im} format. If \code{xi} is in \code{im} format then it is assumed that the pixels will be valued 1 (for foreground), 0 (for background) and NA for unobserved.
If \code{xi} is in \code{owin} format take care to consider what exactly the observation window is (if none is supplied then it will be assumed that the observation window is the smallest rectangle enclosing \code{xi}).}

\item{obswin}{The observation window in \code{owin} format. If it isn't included and \code{xi} is an \code{owin} object then \code{obswin} is taken to be the smallest rectangle enclosing \code{xi}. If \code{xi} is a \code{im} object than \code{obswin} is all the non-NA pixels in \code{xi}.}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window
is smaller than this threshold is given a covariance of NA to avoid instabilities caused by dividing by very small areas,}

\item{modifications}{A list of strings specifying desired modifications or functions to apply to cvchat, cpp1 and phat.
modifications = "all" will select all inbuilt modifications. See details.}

\item{cvchat}{The classical estimate of covariance in \code{im} format. Typically created with \code{racscovariance}.}

\item{cpp1}{Picka's coverage probability estimate in \code{im} format. Typically generated by \code{cppicka} - see help for \code{cppicka} for more information.}

\item{phat}{The classical estimate of coverage probability,
which is the observed area in \code{xi} divided by the total area of the observation window.
See \code{coverageprob} for more information.}

\item{modification}{A string specifying the desired modification. See details.}
}
\value{
\code{racscovariance.cvchat} returns a \pkg{SpatStat} \code{im} object containing the modified estimated covariance.
 The grey scale values in this image represent the covariance for an array of vectors. 
 \code{racscovariances.cvchat} is a named imlist of such image objects for each of the modifications requested.
}
\description{
This function estimate the covariance of a stationary RACS. 
A variety of balanced, partially balanced and classical estimates are available.
}
\details{
Modifies the classical covariance estimator to based on balancing ideas for pair-correlation and centred covariance estimation.
Many of the modifications use Picka's coverage probability estimators in some way.
Modifications available are: 
\itemize{
\item \code{none} Returns cvchat
\item symm Returns the estimated average of cvchat at -v and +v
\item adrian  Similar to Picka additive but uses \code{cpp1} twice and not the refelction of \code{cpp1}
\item mattfeldadd
\item mattfeldmult
\item pickaadd The only modification supplied that provides estimates that satisfy C(v) = 2phat - 1 + Ccomplment(v)
\item pickamult
\item pickahajek
}
}
\section{Functions}{
\itemize{
\item \code{balancedracscovariance.cvchat}: Applies covariance balancing modification to precomputed cvchat, cpp1 and phat

\item \code{balancedracscovariances.cvchat}: Applies multiple modifications simultaneously from a precomputed cvchat, cpp1 and phat

\item \code{cvchats.convolves}: Applies multiple modifications simultaneously from a precomputed convolutions xi*xi, w*w, xi*w and phat
}}

\examples{
xi <- heather$coarse
obswin <- Frame(xi)
balancedcvchats <- balancedracscovariances(xi, obswin = Frame(xi), modifications = "all")
balancedcvchats2 <- byconv.cvchats(xi, obswin = Frame(xi), modifications = "all")

plot.solist(c(balancedcvchats, balancedcvchats2), ncols = 8)
diff <- mapply(function(x, y) x - y, balancedcvchats, balancedcvchats2, SIMPLIFY = FALSE)
plot.solist(diff)

xixi <- setcov(xi, xy = xi)
winwin <- setcov(obswin, xy = xi)
xiwin <- setcov(xi, obswin, xy = xi)
phat <- coverageprob(xi, obswin = Frame(xi))
cvchat <- racscovariance(xi, inclraw = FALSE)
cpp1 <- cppicka(xi, obswin = Frame(heather$coarse))
harmonised <- harmonise.im(cvchat = cvchat, cpp1 = cpp1)
cvchat <- harmonised$cvchat
cpp1 <- harmonised$cpp1

balancedcvchat <- balancedracscovariance.cvchat(cvchat, cpp1, phat, modification = "pickaadd")
balancedcvchats <- balancedracscovariances.cvchat(cvchat, cpp1, phat, modifications = "all")
modifications <- c("none",
 "symm",
 "adrian",
 "mattfeldtadd",
 "mattfeldtmult",
 "pickaadd",
 "pickamult",
 "pickahajek", 
 function(cvchat, cpp1, phat) cvchat)
balancedcvchats <- balancedracscovariances.cvchat(cvchat, cpp1, phat, modifications = modifications)
plot(as.solist(balancedcvchats), equal.ribbon = TRUE)


}
\author{
{Kassel Liam Hingee}
}
\keyword{nonparametric}
\keyword{spatial}
