% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balancedracscovariance_convfirst.R
\name{byconv_cvchats}
\alias{byconv_cvchats}
\alias{cvchats_convolves}
\title{Balanced spatial covariance estimation, also known as `two-point probability', estimator for stationary RACS.}
\usage{
byconv_cvchats(xi, obswin, xy = NULL, setcov_boundarythresh = NULL,
  modifications = "all")

cvchats_convolves(xixi, winwin, xiwin = NULL, phat = NULL,
  modifications = "all")
}
\arguments{
\item{xi}{An observation of the RACS of interest. It can be in \pkg{spatstat}'s \code{owin} or \code{im} format. If \code{xi} is in \code{im} format then it is assumed that the pixels will be valued 1 (for foreground), 0 (for background) and NA for unobserved.
If \code{xi} is in \code{owin} format take care to consider what exactly the observation window is (if none is supplied then it will be assumed that the observation window is the smallest rectangle enclosing \code{xi}).}

\item{obswin}{The observation window in \code{owin} format. If it isn't included and \code{xi} is an \code{owin} object then \code{obswin} is taken to be the smallest rectangle enclosing \code{xi}. If \code{xi} is a \code{im} object than \code{obswin} is all the non-NA pixels in \code{xi}.}

\item{xy}{A raster object that specifies the pixel coordinates of the desired covariance image. In the same vein and as.mask in spatstat. **INCORRECT IT IS PASSED TO OBJECTS BEFORE setcov()}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window
is smaller than this threshold is given a covariance of NA to avoid instabilities caused by dividing by very small areas,}

\item{modifications}{A list of strings specifying desired modifications or functions to apply to cvchat, cpp1 and phat.
modifications = "all" will select all inbuilt modifications. See details.}

\item{xixi}{The convolution of a set representing xi with itself}

\item{winwin}{The convolution of the observation window with itself}

\item{xiwin}{The convolution of the set xi with the observation window}

\item{phat}{The classical estimate of coverage probability,
which is the observed area in \code{xi} divided by the total area of the observation window.
See \code{coverageprob} for more information.}
}
\description{
This function estimate the covariance of a stationary RACS. 
A variety of balanced, partially balanced and classical estimates are available. 
These functions differ to \code{\link{balancedracscovariances}} by computing separately with the convolutions used in the numerator and denominator until end.
}
\section{Functions}{
\itemize{
\item \code{cvchats_convolves}: Applies multiple modifications simultaneously from a precomputed convolutions xi*xi, w*w, xi*w and phat
}}

\examples{
xi <- heather$coarse
obswin <- Frame(xi)
balancedcvchats <- byconv_cvchats(xi, obswin = Frame(xi), modifications = "all")
xixi <- setcov(xi, xy = xi)
winwin <- setcov(obswin, xy = xi)
xiwin <- setcov(xi, obswin, xy = xi)
}
