% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccvc.R
\name{cencovariance}
\alias{cencovariance}
\alias{cencovariance.cvchat}
\title{Centred covariance estimation}
\usage{
cencovariance(xi, obswin = NULL, setcov_boundarythresh = NULL,
  estimators = "all", drop = FALSE)

cencovariance.cvchat(cvchat, cpp1 = NULL, phat = NULL,
  setcov_boundarythresh = NULL, estimators = "all", drop = FALSE)
}
\arguments{
\item{xi}{A binary map of an observation of a RACS of interest. See
\code{\link{stationaryracsinference-package}} for details.}

\item{obswin}{If \code{xi} is an \code{owin} object then \code{obswin} is an
\code{owin} object that specifies the observation window.}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window
is smaller than this threshold is given a covariance of NA to avoid instabilities caused by dividing by very small areas,}

\item{estimators}{A list of strings specifying estimators to use. 
See details.
\code{estimators = "all"} will select all available estimators.}

\item{drop}{If TRUE and one estimator selected then the returned value will be a single \code{im} object and not a list of \code{im} object.}

\item{cvchat}{The traditional estimate of covariance in \code{im} format. 
Typically created with \code{\link{tradcovarest}}.}

\item{cpp1}{Picka's reduced window estimate of coverage probability in \code{im} format - used in improved (balanced) covariance estimators.
Can be generated using \code{\link{cppicka}}.}

\item{phat}{The traditional estimate of coverage probability,
which is the observed foreground area in \code{xi} divided by the total area of the observation window.
See \code{\link{coverageprob}} for more information.}
}
\value{
If \code{drop = TRUE} and a single estimator requested then a
  \code{im} object containing the centred covariance estimate. Otherwise a
  named \code{imlist} of \code{im} objects containing the centred covariance
  estimates for each requested estimator.
}
\description{
This function estimates the centred covariance of a stationary RACS. 
The traditional centred covariance estimator, two 'balanced' estimators suggested by Picka
 and a third 'balanced' estimator inspired by one of Picka's pair-correlation estimators.
}
\details{
The centred covariance of a stationary RACS is \deqn{\kappa(v) =
  C(v) - p^2.}

  The estimators available are (see [Chapter 4, hingee2019thesis] for
  information): 
  \itemize{ 
  \item{\code{trad}} the traditional centred
  covariance estimator 
  \item{\code{mattfeldt}} an estimator inspired by an
  `intrinsically' balanced pair-correlation estimator from Picka that was
  later studied in an isotropic situation by Mattfeldt and Stoyan
  [mattfeldt2000im] 
  \item{\code{pickaint}} Picka's intrinsically' balanced
  centred covariance estimator [picka2000va]. 
  \item{\code{pickaH}} Picka's
  additively' balanced centred covariance estimator [picka2000va].
  }

  Currently computes centred covariance using \code{\link{racscovariance}}.
}
\section{Functions}{
\itemize{
\item \code{cencovariance}: Centred covariance estimates from a binary map.

\item \code{cencovariance.cvchat}: Generates centred covariances estimates from
a traditional estimate of covariance, Picka's reduced window estimate of coverage probability,
and the traditional estimate of coverage probability.
If these estimates already exist then \code{cencovariance.cvchat} can save significant computation time.
}}

\examples{
xi <- heather$coarse
obswin <- Frame(xi)
cencovariance(xi, obswin, estimators = "all")

}
\author{
{Anonymous}
}
\keyword{nonparametric}
\keyword{spatial}
