% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complement.owin.inwindow.R
\name{complement.owin.inwindow}
\alias{complement.owin.inwindow}
\title{The complement of a set within another set}
\usage{
complement.owin.inwindow(x, w)
}
\arguments{
\item{x}{A set in owin format (can be either raster or polygonal).}

\item{w}{An outer window which will contain the complement. Typically an observation window when x is an observed set.}
}
\value{
An owin object.
}
\description{
Primarily for estinating the complement of a set within an observation window.
Given a set \eqn{X} and window \eqn{W} it returns
\deqn{W \cap X^c}
where $X^c$ is the complement of $X$.
}
\details{
The usual complement.owin function in spatstat does not (to my knowledge) accomplish the task that complement.owin.inwindow does as
the spatstat function requires a rectangular outer window.
}
\examples{
xi <- heather$coarse
xic <- complement.owin.inwindow(xi, Frame(heather$coarse))
}
