% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidenceIntervalOfGaussian.R
\name{confidenceIntervalOfGaussian}
\alias{confidenceIntervalOfGaussian}
\title{Confidence intervals for Guassian Distributions}
\usage{
confidenceIntervalOfGaussian(confidenceLevel, sd)
}
\arguments{
\item{confidenceLevel}{The desired confidence level (between 0 and 1)}

\item{sd}{The standard deviation of the Guassian distribution.}
}
\value{
A list of length two: (lower bound, upper bound)
}
\description{
A simple function for getting symmetric confidence intervals for a given confidence level and standard deviation of a Gaussian distribution
}
\examples{
confidenceIntervalOfGaussian(0.95,1)

}

