% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagSphCont.R
\name{contagSphCont}
\alias{contagSphCont}
\title{Disc State Contagion}
\usage{
contagSphCont(xiH, xiHc, p, normalise = FALSE)
}
\arguments{
\item{xiH}{Estimated conditional spherical contact distribution function for \eqn{\Xi}. 
Typically as a \code{fv} object but could also be a vector of values.}

\item{xiHc}{Estimate conditional spherical contact distribution for the complement of \eqn{\Xi}. 
This is called the Conditional Core Probability in Hingee 2016.
Typically is a \code{fv} object.}

\item{p}{An estimate of the coverage fraction of a RACS \eqn{\Xi}.
Typically obtained using \code{coveragefrac}.}

\item{normalise}{Optional. If TRUE \code{contagSphCont} normalises the results so that all RACS return a value between 0 and 1. Default is FALSE.}
}
\value{
An \code{fv} object or a vector the same length as xiH corresponding to the contagion at each r value of xiH
}
\description{
Calculates the disc-state contagion as described in Hingee 2016. It is like the contagion LPI but is based on the spherical contact version of contagion. 
It describes the entropy (mixing) between four possible states of disc
(see Hingee 2016 for more details).
It requires a mixing distance of interest to be chosen by the user
(compared to classical contagion for which this distance is set by the image resolution).
}
\details{
xiH should be a function of radius that estimates the probability 
\deqn{xiH(r)\approx P(B_r(x) \subseteq \Xi^c | x \in \Xi^c)}
 of a disc around an arbitrary point \eqn{x} is contained in \eqn{\Xi^c.}
Similary xiHc should be an estimate of the probability of a disc being fully contained in \eqn{\Xi}
\deqn{xiHc(r)\approx P(B_r(x) \subseteq \Xi | x \in \Xi).}
These can both be obtained using \code{Hest} in \code{spatstat}.

If xiH and xiHc are both fv objects then they must be generated using Hest because the function automatically uses the reduce-sample border correction estimates.
In this case the return value is an fv object.

If \code{normalise} is \code{TRUE} then the result is divided by 
\eqn{\frac{-4}{e}ln(\frac{1}{e})} and added to 1 so that the normalised disc state contagion is between 0 and 1.
}
\examples{
xi <- heather$coarse
obswindow <- Frame(heather$coarse)
p <- coveragefrac(xi,Frame(xi))
xiH <- Hest(xi) #Sph. Contact Distrution Estimate
xiHc <- Hest(complement.owin(xi),obswindow) #Conditional Core Prob. Estimate
plot(xiH,type="l",col="red") 
lines(xiHc,type="l",col="black") 

contagion <- contagSphCont(xiH,xiHc,p)
plot(contagion)

}
\references{
Hingee, K.L. (2016) Statistics for Patch Observations. ISPRS Congress Proceedings p. IPSRS.
}

