% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{covariance}
\alias{covariance}
\title{Covariance estimates, also known as `two-point probabilities' for stationary RACS}
\usage{
covariance(xi, w = NULL, inclraw = FALSE, setCovBoundaryThresh = 0.1 *
  area.owin(w))
}
\arguments{
\item{xi}{An observation (in spatstat owin format or black and white \code{im} object) of the RACS of interest.}

\item{w}{The observation window in \code{owin} format. If itsn't included its taken to be the smallest rectangle enclosing \code{Xi}.}

\item{inclraw}{If TRUE the output will be two \code{im} objects one for the standard estimator and one raw estimate.}

\item{setCovBoundaryThresh}{to avoid instabilities of dividing by very small areas, any vector \eqn{v} set covariance of the boundary smaller than this threshold is given a covariance of NA}
}
\value{
\code{SpatStat} \code{im} objects containing the estimated covariances.
}
\description{
These functions estimate the covariance of a stationary random closed set. 
The covariance is also known as the two-point coverage probability, and very closely related to the semivariogram.
 The covariance of a vector \eqn{v} is the probability of two points separated by a \eqn{v} being covered by \eqn{\Xi}.
}
\details{
The reduced sample estimator is [1]
\eqn{ \hat{C}(v) = \frac{|\Xi \cap W \cap ((\Xi \cap W ) \oplus v)|}{|W \cap W\oplus v|}}
and the raw estimate (if requested) is
\eqn{\hat{\tilde{C}}(v) = \frac{|\Xi \cap W \cap ((\Xi \cap W ) \oplus v)|}{|W|}.}
\code{covariance} uses Fourier transforms to calculate set covariances (using \code{\link[spatstat]{setcov}} function). It is much faster (500 times faster in one comparison) than \code{covarianceMapEst_direct}.
Vectors with small set covariance of the window are eliminated (using \code{setCovBoundaryThresh} because they cause the covariance to be enormous)
}
\examples{
xi <- heather$coarse
covar <- covariance(xi,inclraw=FALSE)
data(balcattapark_coarse)
xi <- balcattapark_coarse$vegmask
w <- balcattapark_coarse$boundary
covar <- covariance(xi,w,inclraw=FALSE)
}
\references{
[1] Molchanov, I. (1997) Statistics of the Boolean Model for Practitioners and Mathematicians. John Wiley & Sons.
}
\author{
{Kassel Hingee}
}
\keyword{nonparametric}
\keyword{spatial}
