% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covEst.R
\name{covarianceEstAtPoint}
\alias{covarianceEstAtPoint}
\alias{covarianceMapEst_direct}
\title{Old covariance estimates}
\usage{
covarianceEstAtPoint(Xi, w, v)

covarianceMapEst_direct(Xi, w, maxXshiftdistance, maxYshiftdistance)
}
\arguments{
\item{Xi}{An observation (in spatstat owin format or for \code{covariance} an \code{im} object is allowed) of the RACS of interest.}

\item{w}{The boundary of the observation in OWIN format. This is needed for the cases where the observation is not a rectangular region.}

\item{v}{A 2D vector in c(x,y) format.}

\item{maxXshiftdistance}{the maximum size of x-component of vectors \eqn{v} to estimate}

\item{maxYshiftdistance}{the maximum size of y-component of vectors \eqn{v} to estimate}
}
\value{
\item{comp1 }{An estimate (assuming stationarity of \eqn{\Xi}) that two points separated by \eqn{v} will be in \eqn{\Xi}.}
\item{comp2 }{Denominator - The set covariance of the boundary \code{w}}
\item{comp3 }{Numerator - The set covariance of \eqn{\Xi_{obs}}}

For \code{covarianceEstAtPoint} these are single numerical values; for \code{covarianceMapEst_direct} they are matrices; for \code{covariance} they are objects of SpatStat's \code{im} class.
}
\description{
These functions estimate the covariance of a stationary random closed set. 
The covariance is also known as the two-point coverage probability, and very closely related to the semivariogram.
 The covariance of a vector \eqn{v} is the probability of two points separated by a \eqn{v} being covered by \eqn{\Xi}.
}
\details{
The reduced sample estimator is [1]
\eqn{ \hat{C}(v) = \frac{|\Xi \cap W \cap ((\Xi \cap W ) \oplus v)|}{|W \cap W\oplus v|}.}
\code{covariance} uses Fourier transforms to calculate set covariances (using \code{\link[spatstat]{setcov}} function). It is much faster (500 times faster in one comparison) than \code{covarianceMapEst_direct}.
Vectors with small set covariance of the window are eliminated (using \code{setCovBoundaryThresh} because they cause the covariance to be enormous)

\code{covarianceEstAtPoint} estimates the covariance of a single vector \eqn{v} by ratioing the set covariance of \code{Xi} to the set covariance of of observation window \code{w}. Set covariance is calculated by intersecting a set with a translated copy of itself.

\code{covarianceMapEst_direct} estimates covariance on a regular grid using the resolution of \code{Xi}. The regular grid extends to \code{maxXshiftdistance} and \code{maxYshiftdistance} in the x and y components respectively. It uses \code{covarianceEstAtPoint} to estimate the covariance at each grid point.
Ignores point estimates that use an area smaller than 10% of the window.
}
\note{
The name of this function is probably a bit confusing. Perhaps call it two-point coverage probability, and the other functions `two-point coverage probability' functions.

Double check that I'm calculating set covariance directly properly (I don't think there is a need to reflect the set, but maybe I got things wrong)
}
\examples{
XiOWIN <- heather$coarse
windowOWIN <- Frame(heather$coarse)

coverageProb <- coveragefrac(XiOWIN,windowOWIN)

covariancePt <- covarianceEstAtPoint(XiOWIN,windowOWIN,c(3,5))

covarianceDirectEst <- covarianceMapEst_direct(XiOWIN,windowOWIN,1,1)
filled.contour(covarianceDirectEst$covariance)

}
\author{
{Kassel Hingee}
}
\keyword{nonparametric}
\keyword{spatial}
