% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageprobabilityEst.R
\name{covpest}
\alias{covpest}
\title{Estimate the coverage fraction of a stationary RACS}
\usage{
covpest(Xi, w)
}
\arguments{
\item{Xi}{An observation of a RACS.}

\item{w}{The window of observation (not necessarily rectangular)}
}
\value{
an estimate of the coverage probability
}
\description{
Estimates the probability of an aribitray point being inside a stationary RACS.
Because the RACS is stationary this is equivalent to estimating \eqn{P(o\in\Xi)}. 
The estimate is simply the fraction of \eqn{\Xi} in the observed window.
}
\details{
in reality the area of Xi is probably slightly different - depends on the sensing method of Xi. Can assume really close though.
}
\author{
Kassel Hingee
}

