% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddotohyperframe.R
\name{ddotohyperframe}
\alias{ddotohyperframe}
\title{trelliscope ddo to spatstat hyperframe conversion}
\usage{
ddotohyperframe(inddo, slots = "all")
}
\arguments{
\item{inddo}{Input ddo object}

\item{slots}{A list of names of slots to extract into a hyperframe - each slot becomes a column in the hyperframe. 
If not supplied then function will try to extract all slots using the first key-value pair as an example.}
}
\value{
A hyperframe, each column corresponding to a slot.
}
\description{
Extracts data from tessera's ddo objects and puts them into a hyperframe.
}
\examples{
applyestimators <- function(obs,w){
#coverage fraction
coveragefrac <- coveragefrac(obs,w)
#covariance
covariances <- covariance(obs,w)

#Sph Contact Functions. 
sphContact <- list(xiHest = Hest(obs),
                   notxiHest = Hest(complement.owin(obs))) 

#contagion estimates
sphContactHarm <-  harmonise(xiHest = sphContact[[1]],notxiHest = sphContact[[2]])
sphcontactcontag <- list(
  contagionEst = contagSphCont(sphContactHarm$xiHest$km,
                               sphContactHarm$notxiHest$km,
                               coveragefrac),
  r = sphContactHarm$xiHest$r)

return(list(
  coveragefrac = coveragefrac,
  covariance = covariances$covariance,
  sphcontactxi = sphContact$xiHest,
  sphcontactnotxi = sphContact$notxiHest,
  sphcontactcontag = sphcontactcontag
))
}

estimates <- datadr::kvPairs(
     datadr::kvPair("medium",applyestimators(heather$medium,Frame(heather$medium))),
     datadr::kvPair("coarse",applyestimators(heather$coarse,Frame(heather$coarse)))
     )
estimatesddo <- datadr::ddo(estimates)

hframe <- ddotohyperframe(estimatesddo,slots = c("coveragefrac","sphcontactxi"))
hframe <- ddotohyperframe(estimatesddo)
plot(hframe[,"sphcontactxi",drop=TRUE], main = "Spherical Contact Distribution:\\n Diggle's Heather Data")

}

