% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varofobsarea.R
\name{expectedarea}
\alias{expectedarea}
\alias{varofobsarea.v1}
\title{Variance Estimates for Observed Area}
\usage{
expectedarea(xi, obswin, p01 = NA, p10 = NA)

varofobsarea.v1(xi, obswin, corrrad, corrstepheight, p01, p10)
}
\arguments{
\item{xi}{An observation of the RACS of interest in 1, 0, or NA valued pixels.
Pixels must be square.}

\item{obswin}{Observation window}

\item{p01}{Probability of an randomly chosen fallibly-classified
tree pixel is not in the true tree canopy}

\item{p10}{Probability that a falliblty classified non-tree pixel is 
in the tree canopy.}

\item{corrrad}{Radius of the step function in the correlation}

\item{corrstepheight}{Height of the step in the correlation}
}
\description{
Estimates the variance of the area of a cover type observed in a thematic map created using a fallible classifier from remote sensing.
}
\details{
To install OpenImageR had to install libtiff5-dev on my ubuntu machine
}
\section{Functions}{
\itemize{
\item \code{varofobsarea.v1}: The variance of the expected area.
}}

\examples{
xi <- as.im(heather$coarse, na.replace = 0)
obswin <- Frame(xi)
corrrad <- 10 #10 pixels
corrstepheight <- 0.9
p01 <- 0.05
p10 <- 0.01
xi.sum <- sum(xi)
exparea <- expectedarea(xi, obswin, p01 = p01, p10 = p10)
varguess <- varofobsarea.v1(xi, obswin, corrrad, corrstepheight, p01, p10)
stdofareaest <- sqrt(varguess)
}
\author{
{Kassel Hingee}
}
