% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbl.R
\name{gbl}
\alias{gbl}
\alias{gbl.cvchat}
\title{Gliding box lacunarity estimatation using all estimators}
\usage{
gbl(xi, boxwidths, estimators = c("GBLg.mattfeldt", "GBLg.pickaint",
  "GBLg.pickaH", "GBLcc.mattfeldt", "GBLcc.pickaint", "GBLc", "GBLgb"),
  obswin = NULL, includenormed = FALSE,
  setcov_boundarythresh = 1e-06)

gbl.cvchat(boxwidths, estimators = c("GBLg.mattfeldt", "GBLg.pickaint",
  "GBLg.pickaH", "GBLcc.mattfeldt", "GBLcc.pickaint", "GBLc"),
  phat = NULL, cvchat = NULL, cpp1 = NULL)
}
\arguments{
\item{xi}{A binary map of an observation of a RACS of interest. See
\code{\link{stationaryracsinference-package}} for details.}

\item{boxwidths}{A list of box widths}

\item{estimators}{A list of estimator names - see details for possibilities. \code{estimators = "all"} will select all estimators.}

\item{obswin}{If \code{xi} is an \code{owin} object then \code{obswin} is an
\code{owin} object that specifies the observation window.}

\item{includenormed}{A logical value. If TRUE then GBL estimates normalised by the GBL values at zero will be included in a returned list of fv objects}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window is smaller than this threshold
is given a covariance estimate (and other similar estimate) of NA to avoid instabilities caused by dividing by very small areas.
If NULL is supplied (default) then 1E-6 is used.}

\item{phat}{Traditional estimate of coverage probability.}

\item{cvchat}{Traditional estimate of covariance (often from \code{tradcovarest}).}

\item{cpp1}{Picka's estimate of coverage probability (often from \code{cppicka}).}
}
\value{
An \code{fv} object.
}
\description{
Estimates gliding box lacunarity (GBL) using all estimators described in (Hingee et al., 2017) from binary maps for square boxes.
It calls the functions \code{gblc}, \code{gblg}, \code{gblcc} and \code{gbltrad}.
}
\details{
The function is not able to estimate GBL for non-square boxes as the gliding box estimator is included.
To estimate GBL for non-square boxes use \code{gblcc} or \code{gblg} directly.

If \code{xi} is in \code{owin} format then \code{obswin} and \code{xi} are converted
 into a binary map in \code{im} format using \code{\link[spatstat]{as.im}}

The estimators available are
\itemize{
\item{\code{"GBLc"}} The unmodified (unbalanced) covariance estimator provided by \code{\link{gblc}}
\item{\code{"GBLgb"}} The Gliding-Box estimator of Allain and Cloitre (1991). Calls \code{\link{gbltrad}}
\item{\code{"GBLg.mattfeldt"}} See help for \code{\link{gblg}}
\item{\code{"GBLg.pickaint"}} See help for \code{\link{gblg}}
\item{\code{"GBLg.pickaH"}} See help for \code{\link{gblg}}
\item{\code{"GBLcc.mattfeldt"}} See help for \code{\link{gblcc}}
\item{\code{"GBLcc.pickaint"}} See help for \code{\link{gblcc}}
\item{\code{"GBLcc.pickaH"}} See help for \code{\link{gblcc}}
}
}
\section{Functions}{
\itemize{
\item \code{gbl}: Computes GBL estimates from a binary map.

\item \code{gbl.cvchat}: Computes covariance-based estimator of GBL from the traditional estimate of covariance,
Picka's reduced window coverage probability estimates and the traditional coverage probability estimate.
}}

\examples{
xi <- heather$coarse
xi <- as.im(xi, value = TRUE, na.replace = FALSE)
gblests <- gbl(xi, seq(1, 10, by = 0.1))
}
\references{
Allain, C. and Cloitre, M. (1991) Characterizing the lacunarity of random and deterministic fractal sets. \emph{Physical Review A}, 44, 3552-3558.

Hingee K, Baddeley A, Caccetta P, Nair G (2017). Computation of lacunarity from covariance of spatial binary maps. \emph{Journal of Agricultural, Biological and Environmental Statistics}. Submitted.
}
