% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gblcc.R
\name{gblcc}
\alias{gblcc}
\title{Centred covariance based estimates of gliding box lacunarity}
\usage{
gblcc(boxes, cencovar = NULL, p = NULL, xiim = NULL,
  estimator = "pickaH")
}
\arguments{
\item{boxes}{Either a list of sidelengths for square boxes or a list of \code{owin} objects of any shape.}

\item{cencovar}{A \code{im} object containing the centred covariance function}

\item{p}{The coverage probability. Typically estimated by the fraction of the observation window covered by the set of interest.}

\item{xiim}{An observation of a stationary RACS in \code{im} format. \code{xiim} must have values of either 1, 0 or NA; 1 denotes inside the RACS, 0 denotes outside, and NA denotes unobserved.}

\item{estimator}{If an observation \code{xiim} is passed then \code{estimator} will select the balancing method that \code{ccvc} uses to estimate the centred covariance.}
}
\value{
If \code{boxes} is a list of numerical values then GBL is estimated for square boxes with side length given by \code{boxes}.
 The returned object is then an \code{fv} object containing estimates of GBL, box mass variance and box mass mean.
 
 If \code{boxes} is a list of owin objects then \code{gblcc} returns a dataframe of with columns corresponding to estimates of GBL, box mass variance and box mass mean.
 Note if NA or NaN values in the \code{covariance} object are used then \code{gblc} will return NA or NaN instead of an GBL value.
}
\description{
Estimates the gliding box lacunarity (GBL) of a stationary RACS using centred covariance estimates (Hingee et al., 2017).
 The centred covariance and coverage probability can be provided or estimated from binary map.
}
\details{
If we denoted the estimated centred covariance by
\eqn{\hat{\kappa}(v)}{k(v)} and coverage probability \eqn{\hat{p}}{p} then
the estimate of GBL is
\deqn{1 + \frac{1}{\hat{p}^2 |B|^2}\int \gamma_B(v)\hat{\kappa}(v)dv }{1 + \int gammaB(v) k(v) dv  /  (p^2 |B|^2).}
}
\examples{
xi <- heather$coarse
cencovar <- cencovariance(xi, obswin = Frame(xi), estimators = c("pickaH"), drop = TRUE)
p <- area(xi) / area(Frame(xi))
if (interactive()) {
sidelengths <- seq(0.3, 14, by = 0.2)
} else {
sidelengths <- seq(0.3, 14, by = 1)
}
gblccest <- gblcc(sidelengths, cencovar, p)
# what is the GBL estimates for boxes that are discs?
discboxes <- lapply(sidelengths / 2, disc)
discgbls <- gblcc(discboxes, cencovar, p)
# points(sidelengths, discgbls)

#direct to an image
xiim <- as.im(xi, na.replace = 0)
gblccest <- gblcc(sidelengths, xiim = xiim, estimator = "pickaH")

}
\references{
Hingee K, Baddeley A, Caccetta P, Nair G (2017). Computation of lacunarity from covariance of spatial binary maps. \emph{Journal of Agricultural, Biological and Environmental Statistics}. Submitted.
}
\keyword{nonparametric}
\keyword{spatial}
