% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innerprodim.R
\name{innerprod.im}
\alias{innerprod.im}
\title{Inner Product of Two Functions Represented as Images}
\usage{
innerprod.im(A, B, outsideA = NA, outsideB = NA, na.rm = FALSE)
}
\arguments{
\item{A}{An \code{im} object containing function values.}

\item{B}{An \code{im} object.}

\item{outsideA}{The value of the function represented by \code{A} outside the domain of \code{A}. Typically will be 0 or NA. Default is NA.}

\item{outsideB}{The value of the function represented by \code{B} outside the domain of \code{B}. Typically will be 0 or NA. Default is NA.}

\item{na.rm}{Logical. If TRUE NA values are skipped in the summation of the product of the images.}
}
\value{
If the inner product can be computed then returns sum(A*B), otherwise returns NA.
}
\description{
Given two functions, f and g, that map from 2D space to 1D, and values of f and g represented as \code{im} objects
 then \code{innerprod.im} computes the (function space) inner product
\deqn{\int f(v) g(v) dv.}
}
\details{
Harmonises the two input images and computes their multiple. The sum of values of their multiples is then returned.
outsideA and outsideB are used to determine how to act if the inner product requires values outside the domain of A and B.
For example if outsideA=0 and the domain of B is larger than A's domain then the inner product can still be computed.
However if A is NA outside (e.g. not known/not provided) and the domain of B is larger than A's domain then the inner product is not computable.
}
\examples{
xi <- heather$coarse
covar <- racscovariance(xi, inclraw = FALSE)
B <- setcov(square(1))
innerprod.im(covar, B, outsideB = 0)
}
