% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratebyrad.R
\name{integratebyradius}
\alias{integratebyradius}
\title{Estimates the integral of image values within a radius r.}
\usage{
integratebyradius(centre, values)
}
\arguments{
\item{centre}{The centre location}

\item{values}{A pixel image to be integrated}
}
\value{
An fv object
}
\description{
Estimates the integral of image values within a radius r from the centre.
}
\examples{
twptprob <- covariance(heather$coarse,Frame(heather$coarse))
kfcn <- integratebyradius(c(0,0),twptprob)
plot(kfcn)


}

