% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isbinarymap.R
\name{isbinarymap}
\alias{isbinarymap}
\title{Test if an im object is a binary map}
\usage{
isbinarymap(xiim, requiretrue = FALSE)
}
\arguments{
\item{xiim}{an image object}

\item{requiretrue}{Logical. If TRUE then isbinarymap will error if xiim is NOT a binary map.}
}
\value{
Logical value. TRUE if \code{xiim} is a binary map. Otherwise FALSE.
}
\description{
Tests \code{xiim} to see if it is a binary map. The pixel values must be of logical type (TRUE, FALSE and NA only), or numerical with values of 0, 1 and NA.
}
\examples{
#The following return TRUE
isbinarymap(as.im(heather$coarse, na.value = 0))
isbinarymap(as.im(heather$coarse, na.value = FALSE, value = TRUE))
#the following return FALSE
isbinarymap(as.im(heather$coarse, na.value = 0.2, value = 1))
isbinarymap(as.im(heather$coarse, na.value = 0, value = 1.5))
}
