% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernelsmooth.R
\name{kernelsmooth}
\alias{kernelsmooth}
\title{Kernel Smoothing Tools}
\usage{
kernelsmooth(im, bandwidth, kernel = "Epanechnikov")
}
\arguments{
\item{im}{an image for smoothing.}

\item{bandwidth}{Bandwidth for the chosen kernel (in reciprocal of the units of \code{im})}

\item{kernel}{Specifies the kernel to use. Currently only Epanechnikov is supported (and is the default).}
}
\value{
A smoothed image in spatstat \code{im} format
}
\description{
Function for kernel smoothing spatstat images. This is mostly for smoothing spectral density estimates.
Only kernel currently supported is Epanechnikov.
}
\details{
Uses FFT to do the calculation.
}
\section{Notes}{
 This function is almost superfluous - \code{spatstat} has a function \code{Smooth.im} for smoothing using Gaussian kernels.
}

\examples{
#smooth a step image
stepim <- im(matrix(0,nrow=100,ncol=100),xrange=c(-1,1),yrange=c(-1,1))
stepim[owin(xrange=c(-0.5,0.5),yrange=c(-0.5,0.5))] <- 1
# plot(stepim,axes=TRUE)

smstepim <- kernelsmooth(stepim,0.5)
# plot(smstepim,axes=TRUE)
}
