% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lac_gb.R
\name{lacgb}
\alias{lacgb}
\title{Gliding Box lacunarity from a black and white image}
\usage{
lacgb(img, bandwidths)
}
\arguments{
\item{img}{An image of 0's and 1's.}

\item{bandwidths}{A list of bandwidths (half the box sidelengths) in the dimensions of \code{img}}
}
\value{
An fv object with items for no edge correction and reduced sample border correction
}
\description{
Calculates the gliding box lacunarity
}
\details{
Calculates the gliding box lacunarity for a given range of box sizes (`radius').
The centres are given the pixels of \code{img}.

Note: (1) The bandwidths are rounded such that box sidelengths are an odd number of pixels across. (2) The reduced sample points are given by erosion of the image by a disc with radius bandwidth+0.5*pixelwidth, which is note quite erosion by a square.
}
\examples{
data(balcattapark_coarse)
img <- as.im(balcattapark_coarse$vegmask)
bandwidths <- c(1.6,1.9,3.2,5*0.8) #in units of img
lac <- lacgb(img,bandwidths)
plot(lac, cbind(RS,nobord) ~ b)

}
\references{
The gliding box algorithm is described in Allain, C. and Cloitre, M. (1991) Characterizing the lacunarity of random and deterministic fractal sets. Physical Review A, 44, 3552-3558.
}

