% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manyfvevalat.R
\name{manyfvevalat}
\alias{manyfvevalat}
\title{Evaluate a list of fv objects at a particular argument}
\usage{
manyfvevalat(fvlist, argval, value = ".y", extrapolate = TRUE)
}
\arguments{
\item{fvlist}{A list of fv objects}

\item{argval}{The value of argument that each function is evaluated at.}

\item{value}{The function value to use, ".y" uses the default provided by the fv objects.
Like \code{value} in \code{\link[spatstat]{as.function.fv}} but in this case can only only be a single string.}

\item{extrapolate}{From \code{\link[spatstat]{as.function.fv}}, whether to extrapolate the fv objects when the provided argval is outside their range.}
}
\value{
A list of values of each fv object at \code{argval}.
}
\description{
Given a list of fv objects evaluates each at the given argument, and returns a list of the same length.
}
\examples{
fv1 <- Hest(heather$coarse)
fv2 <- Hest(complement.owin(heather$coarse))
fvlist <- list(scdxi = fv1, scdnotxi = fv2) 
manyfvevalat(fvlist,0.19)
# out <- manylines.fv(fvlist, fmla="km~r")
# abline(v=0.19)
}
