% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmanyfv.R
\name{manylines.fv}
\alias{manylines.fv}
\title{Plot multiple fv objects of similar ilk}
\usage{
manylines.fv(fvlist, fmla = NULL, ..., add = FALSE, xlim = NULL,
  ylim = NULL)
}
\arguments{
\item{fvlist}{A list of fv objects}

\item{fmla}{an R language formula determining which variables or expressions are plotted.
Either a formula object, or a string that can be parsed as a formula.
It is passed to \code{plot.fv} - see \code{\link{plot.fv}} for details.}

\item{...}{Plotting parameters to be passed to \code{plot.window}, \code{title} and \code{\link{plot.fv}}}

\item{add}{If true will add lines to current plot, otherwise will start new plot and will require xlim and ylim.}

\item{xlim}{Plotting limits of the horizontal direction}

\item{ylim}{Plotting limits of the vertical direction}
}
\value{
List of return values from plot.fv
}
\description{
Plots multiple fv objects on the same window.
}
\examples{
hestf <- Hest(heather$fine, W =Frame(heather$fine))
hestm <- Hest(heather$medium, W =Frame(heather$medium))
hestc <- Hest(heather$coarse, W =Frame(heather$coarse))
fvlist <- list(hestf,hestm,hestc)
names(fvlist) <- c("fine", "medium", "coarse")
manylines.fv(fvlist, km~r)

## Fancier plot
out <- manylines.fv(fvlist,  
main="Spherical Contact Distribution\\n from different resolutions", col=rainbow(length(fvlist)))
}

