% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medianise.R
\name{median_ise.fvlist}
\alias{median_ise.fvlist}
\title{Median Integrated Squared Error of a List of fv Objects}
\usage{
median_ise.fvlist(object, benchfv, domainlim, equiv = NULL,
  avoverdomain = FALSE, fixeddomain = FALSE, acceptableISEerrorrate = 0.1,
  ...)
}
\arguments{
\item{object}{List of function objects to compare to benchfv}

\item{benchfv}{An fv function object to treat as the true or reference values.}

\item{domainlim}{A list of two entries giving the lower and upper bound of the domain to integrate over}

\item{equiv}{Passed to eval.fv - see help for eval.fv for more details.
It is gives the functional value names to map to each other when comparing to benchfv. Default is NULL.}

\item{avoverdomain}{Divide the integral by the length of the domain (i.e. divide by \eqn{b_i} - \eqn{a_i} for each \eqn{\hat{f}_i}}

\item{fixeddomain}{If the domain of the function \eqn{\hat{f}_i}
is a non-equal subset of the domain given by \code{domainlim} then set the integrated error to NA.}

\item{acceptableISEerrorrate}{The number of bad integrations that it is ok to ignore when computing the median ISE.}

\item{...}{Arguments passed to integrate().}
}
\value{
A numeric value that is the Median Integrated Squared Error - see Details.
}
\description{
Computes the median integrated squared error (defined below) of a list of fv objects given reference fv object.
}
\details{
We define the median integrated squared error of a collection of estimates of functions \eqn{\hat{f}_i}
and a reference function \eqn{f} as 
\deqn{
\left\{\int_{a_i}^{b_i} \left(\hat{f}_i(s) - f(s) \right)^2 ds  : i = 1, 2, 3, .... , n \right\},
}
where the domain of integration $[a_i, b_i]$ is 
the largest possible interval between \code{domainlim[[1]]} and \code{domainlim[[2]]} for which the function
\eqn{\hat{f}_i} exists.
 (Unless \code{fixeddomain} is \code{TRUE} in which case the domain of integration is fixed to \code{domainlim})

The above integral is computed using the numerical integration function \code{integrate} and
 if this numerical integration fails then the result is ignored in the final computation of the median.
\code{acceptableISEerrorrate} can be used to set the proportion of these failures that is acceptable for calculating the median.

The function fails if there is y-value name of the reference fv object
 is equal to a y-value name in the list fv objects that that you don't want to compare to
  (e.g. if the list of fv objects also contain the reference value).
}
\examples{
obspatterns <- rpoispp(10, nsim = 10)
object <- lapply(obspatterns, Kest, W = Frame(obspatterns[[1]]))
benchfv <- as.fv(object[[1]][, c("r", "theo"), drop = TRUE])
names(benchfv) <- c("r", "bench")
median_ise.fvlist(object, benchfv, c(0.1, 0.2), equiv = list(bench = "iso"), avoverdomain = TRUE)

}
\author{
Kassel Hingee
}
