% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlfromshapeandraster.R
\name{MVLests_files}
\alias{MVLests_files}
\alias{MVLest_multipleregions}
\alias{MVLest_region}
\alias{converttologicalim}
\alias{plot_MVLest_region}
\alias{plot_MVLest_allregions}
\title{MVL estimates from shapefiles, SpatialPolygonDataFrames  and raster data}
\usage{
MVLests_files(shapefile, rasterfile, frange, NArange, boxwidths,
  estimators = c("MVLg.mattfeldt", "MVLg.pickaint", "MVLcc.mattfeldt",
  "MVLcc.pickaint", "MVLcc.pickaH", "MVLc", "MVLgb"),
  normalisebyMVLzero = TRUE, display = TRUE)

MVLest_multipleregions(polysdf, rasterlayer, frange, NArange, boxwidths,
  estimators = c("MVLg.mattfeldt", "MVLg.pickaint", "MVLcc.mattfeldt",
  "MVLcc.pickaint", "MVLcc.pickaH", "MVLc", "MVLgb"),
  normalisebyMVLzero = FALSE, display = TRUE)

MVLest_region(polydf, rasterlayer, frange, NArange, boxwidths,
  estimators = c("MVLg.mattfeldt", "MVLg.pickaint", "MVLcc.mattfeldt",
  "MVLcc.pickaint", "MVLcc.pickaH", "MVLc", "MVLgb"),
  normalisebyMVLzero = FALSE, showprogress = FALSE)

converttologicalim(polydf, rasterlayer, frange, NArange)

plot_MVLest_region(returnedlist, plot.im.args = list(main = "Class Image",
  axes = TRUE), plot.mvl.args = NULL)

plot_MVLest_allregions(returnedlist, normaliseAtStart = FALSE,
  estname = "mvlcc.pickaH", ...)
}
\arguments{
\item{shapefile}{The path to a shapefile}

\item{rasterfile}{The path to a single layer raster file readable to RGDAL}

\item{frange}{is list of length 2 specifying the minimum and maximum pixel values to assign as foreground. The range is inclusive.}

\item{NArange}{is list of length 2 specifying the minimum and maximum pixel values to assign as NA values. The range is inclusive and overrides any assignment by frange.}

\item{boxwidths}{The set of widths of square boxes to estimate MVL for.}

\item{estimators}{A list of names of MVL estimators to use. See \code{mvl()} for available list and more information}

\item{normalisebyMVLzero}{Logical. If TRUE the returned MVL estimates will be normalised by the MVL in the limit of very small boxes,
which is \eqn{p (1-p)/p^2}, where \eqn{p} is the coverage probability.}

\item{display}{If TRUE then \code{plot_MVLest_region} is called so that the results are automatically plotted}

\item{polysdf}{A SpatialPolygonsDataFrame with each feature corresponding to a different region.}

\item{rasterlayer}{is a raster layer object}

\item{polydf}{is a SpatialPolygonsDataFrame with a single region}

\item{showprogress}{If TRUE then updates on progress will be printed to the screen.}

\item{returnedlist}{The results of a call to \code{MVLest_region}}

\item{plot.im.args}{A named list of arguments passed to \code{plot.im} for plotting the class image.}

\item{plot.mvl.args}{A names list of argument passed to \code{plot.fv} for plotting the MVL estimates.}
}
\description{
Functions for estimating MVL from regions specified in shapefiles or SpatialPolygonDataFrames given a raster file, or a \code{RasterLayer} object.
}
\section{Functions}{
\itemize{
\item \code{MVLests_files}: Returns MVL estimates from regions specified in shapefile using raster data given in rasterfile

\item \code{MVLest_multipleregions}: Splits a SpatialPolygonDataFrame into component features and estimates MVL using \code{rasterlayer} data in each of these features.

\item \code{MVLest_region}: Computes MVL estimates assuming that SpatialPolygonsDataFrame (polydf) representing a single region. 
Returns a list of the three items (1) the logical image used, (2) the MVL estimates and (3) the polydf used an input.

\item \code{converttologicalim}: Converts data contained in \code{rasterlayer} into a \code{spatstat im} object with logical values of TRUE (for foreground), FALSE (for background) and NA.

\item \code{plot_MVLest_region}: A special plotting function for the results of \code{MVLest_region}. Plots the image used, the MVL estimates and includes polygon attributed below plots.

\item \code{plot_MVLest_allregions}: A speical plotting function for the results of \code{MVLest_multipleregions} and \code{MVLest_files}.
}}

