% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgb.R
\name{mvlgb}
\alias{mvlgb}
\alias{lacgb}
\title{Gliding Box estimation of mass variance lacunarity}
\usage{
mvlgb(img, sidelengths, inclraw = FALSE, W = Frame(img))

lacgb(img, sidelengths, inclraw = FALSE, W = Frame(img))
}
\arguments{
\item{img}{An image of pixels valued either \code{0}, \code{1} or \code{NA}. \code{NA} valued pixels are assumed to be outside the observation window.}

\item{sidelengths}{A list of suggested box side lengths in the same units as \code{img}. Note the actual side lengths used will be the closest multiple of an odd number of pixel widths.}

\item{inclraw}{If TRUE the function will also return a gliding box estimate that ignores edge effects. (Default is FALSE)}

\item{W}{Optional observation window. The observation window used for the estimator will be the intersection of \code{W} and the pixels that are not \code{NA} in \code{img}.}
}
\value{
An \code{fv} object with a column called \code{MVL} for the usual gliding box estimate described in [1]. 
 The side lengths (labelled \code{s}) are always odd multiples of the pixel width.
 
Another column, called \code{raw}, is included if \code{inclraw=TRUE}. 
This column represents a version of the gliding box algorithm that pretends that the items/cover of interest are fully observed
 (i.e. that the observation window is the entire plane).
}
\description{
Calculates the gliding box estimate [1] of mass variance lacunarity from a bi-tonal image.
}
\details{
Calculates the gliding box estimate [1] of mass variance lacunarity for a given range of square box sizes. 
The algorithm uses the pixel locations in \code{img} as an array of box centre locations to compute
 the mean and variance of the area in a random box of a given size.
Locations where the box is not completely within the observation window are ignored.
 
WARNING: This function needs the \code{roll_sum} function in \code{RcppRoll} to operate.

Note: The sidelengths are rounded such that they are an odd number of pixels across.
}
\examples{
img <- as.im(heather$coarse,na.replace=0)
sidelengths <- seq(0.2,14,by=0.2) #in units of img
lac <- mvlgb(img,sidelengths, inclraw=TRUE)
plot(lac, cbind(MVL,raw) ~ s)

}
\references{
[1] Allain, C. and Cloitre, M. (1991) Characterizing the lacunarity of random and deterministic fractal sets. Physical Review A, 44, 3552-3558.
}
