% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcln.R
\name{pcln}
\alias{pcln}
\alias{pcln.cvchat}
\title{Balanced estimation of centred spatial covariance for stationary RACS}
\usage{
pcln(xi, obswin, setcov_boundarythresh = NULL, modifications = NULL)

pcln.cvchat(cvchat, cpp1 = NULL, phat = NULL, modifications = NULL)
}
\arguments{
\item{xi}{An observation of the RACS of interest. It can be in \pkg{spatstat}'s \code{owin} or \code{im} format. If \code{xi} is in \code{im} format then it is assumed that the pixels will be valued 1 (for foreground), 0 (for background) and NA for unobserved.
If \code{xi} is in \code{owin} format take care to consider what exactly the observation window is (if none is supplied then it will be assumed that the observation window is the smallest rectangle enclosing \code{xi}).}

\item{obswin}{The observation window in \code{owin} format. If it isn't included and \code{xi} is an \code{owin} object then \code{obswin} is taken to be the smallest rectangle enclosing \code{xi}. If \code{xi} is a \code{im} object than \code{obswin} is all the non-NA pixels in \code{xi}.}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window
is smaller than this threshold is given a covariance of NA to avoid instabilities caused by dividing by very small areas,}

\item{modifications}{A list of strings specifying desired modifications or functions to apply to cvchat, cpp1 and phat.
modifications = "all" will select all inbuilt modifications. See details.}

\item{cvchat}{The classical estimate of covariance in \code{im} format. Typically created with \code{racscovariance}.}

\item{cpp1}{Picka's coverage probability estimate in \code{im} format. Typically generated by \code{cppicka} - see help for \code{cppicka} for more information.}

\item{phat}{The classical estimate of coverage probability,
which is the observed area in \code{xi} divided by the total area of the observation window.
See \code{coverageprob} for more information.}
}
\value{
A named imlist of \pkg{SpatStat} \code{im} objects containing the centred covariance estimates from the different modifications.
}
\description{
Estimates the pair-correlation function of a stationary RACS. 
A variety of balanced, partially balanced and classical estimates are available.
}
\details{
Modifications available are: 
\itemize{
\item \code{none} Returns cvchat / (phat * phat)
\item symm Returns the estimated average of cvchat at -v and +v, divided by (phat * phat)
\item mattfeld The pair-correlation estimator from Mattfeldt's paper
\item pickaint Picka's intrinsic modification which modifies the coverage probability estimators
\item pickahajek Picka's Hajek-based modification which additively modifies the covariance estimate.
}
}
\section{Functions}{
\itemize{
\item \code{pcln.cvchat}: Applies multiple modifications simultaneously from a precomputed cvchat, cpp1 and phat
}}

\examples{
xi <- heather$coarse
obswin <- Frame(xi)
phat <- coverageprob(xi, obswin = Frame(xi))
cvchat <- racscovariance(xi, inclraw = FALSE)
cpp1 <- cppicka(xi, obswin = Frame(heather$coarse))

pclns <- pcln.cvchat(cvchat, cpp1, phat, modifications = "all")

}
\author{
{Kassel Liam Hingee}
}
\keyword{nonparametric}
\keyword{spatial}
