% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbooldeterministicdiscs.R
\name{rbdd}
\alias{rbdd}
\alias{bddcoverageprob}
\alias{bdd_covar.iso}
\alias{bdd.covar}
\alias{bdd.specdens}
\alias{bdd.specdensAtOrigin}
\title{Simulation of Boolean Model of Deterministic Discs}
\usage{
rbdd(lambda, discr, window, seed = NULL)

bddcoverageprob(lambda, discr)

bdd_covar.iso(r, lambda, discr)

bdd.covar(xrange, yrange, eps, lambda, discr)

bdd.specdens(lambda, discr)

bdd.specdensAtOrigin(lambda, discr)
}
\arguments{
\item{lambda}{Intensity of the germ process (which is a Poisson point process)}

\item{discr}{Radius of the discs}

\item{window}{The window to simulate in (an owin object)}

\item{seed}{Optional input (default in NULL). Is an integer passed to \code{\link{base}{set.seed}}. Used to reproduce patterns exactly.}

\item{r}{is the radius to calculate covariance}

\item{xrange}{range of x values for \code{bdd.covar}}

\item{yrange}{range of y values for \code{bdd.covar}}

\item{eps}{list of length 2 of the steps between samples points in x and y respectively for \code{bdd.covar}.}
}
\value{
Depends on the function used (see Functions section).
}
\description{
Functions for simulating a Boolean model with grains that are discs of fixed constant radius (the abreviation bdd is short for Boolean model with Determinisitic Discs).
A Boolean model is a two stage model, first the locations of grains, the germs are randomly distributed according to a Poisson point process, then a random grain is placed on each germ independently.
A thorough introduction can be found in [3].
Also described in this help file are functions for calculating from model parameters the true coverage probability, covariance, and spectral density [1,2].
}
\section{Functions}{
\itemize{
\item \code{rbdd}: Returns an \code{owin} that is a set generated by simulating a deterministic disc Boolean
 model with specified intensity and disc radius.
 The window information is not contained in this object.
 If the simulated set is empy then an empty \code{owin} object is returned.
The point process of germs is generated using spatstat's \code{\link[spatstat]{rpoispp}}.

\item \code{bddcoverageprob}: Returns the true coverage probability given the intensity and disc radius.

\item \code{bdd_covar.iso}: Returns the true covariance of points separated by a distance \code{r} given the intensity, \code{lambda} and disc radius \code{discr} of the model.

\item \code{bdd.covar}: Returns an image of the covariance as calculated from disc radius and intensity.

\item \code{bdd.specdens}: Computes the spectral density using the theoretical covariance and FFT

\item \code{bdd.specdensAtOrigin}: Calculates spectral density at at the origin using the theoretical covariance.
It is the integral of the (covariance - (coverage probability)^2) over all space.
}}

\section{WARNING}{

The returned object of \code{rbdd} is only the contents of Xi and thus could be much smaller than the window (e.g. when the simulated set is empty).
}

\examples{
#Boolean model with discs of radius 10.
#The intensity has been chosen such that the true coverage probability is very close to 0.5.
discr <- 10
w <- owin(xrange=c(0,100),c(0,100))
lambda <- 2.2064E-3
xi <- rbdd(lambda,discr,w)
plot(xi)
plot(w,add=TRUE)

#calculate theoretical values of the model
truecoverageprob <- bddcoverageprob(lambda,discr)
truecovariance <- bdd.covar(
                   c(-10,10),c(-10,10),c(0.2,0.2),lambda,discr)
thspecdens <- bdd.specdens(lambda,discr)
thspecdens_origin <- bdd.specdensAtOrigin(lambda,discr)
thspecdens[round(dim(thspecdens)[2]/2),round(dim(thspecdens)[1]/2)]
}
\references{
[1] Koch, K., Ohser, J. and Schladitz, K. (2003) Spectral Theory for Random Closed Sets and Estimating the Covariance via Frequency Space. Advances in Applied Probability, 35, 603-613.

[2] Bohm, S., Heinrich, L. and Schmidt, V. (2004) Kernel Estimation of the Spectral Density of Stationary Random Closed Sets. Australian and New Zealand Journal of Statistics, 46, 41-51.

[3] Chiu, S.N., Stoyan, D., Kendall, W.S. and Mecke, J. (2013) Stochastic Geometry and Its Applications, 3rd ed. Chichester, United Kingdom: John Wiley & Sons.
}
\keyword{datagen}
\keyword{spatial}
