% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rboollnormdiscs.R
\name{rboollognormdiscs}
\alias{rboollognormdiscs}
\title{Simulate a Boolean model of discs with log normal disc radii and Poisson germs}
\usage{
rboollognormdiscs(window, bufferdist, lambda, meanlog, sdlog, seed = NULL)
}
\arguments{
\item{window}{An \code{owin} object specifying the desired simulation region}

\item{bufferdist}{A distance to expand \code{window} so that discs with centres near \code{window} are also simulated.}

\item{lambda}{Intensity of the Poisson point process, passed to \code{\link{spatstat}{rpoispp}}.  It could be either a single positive number, a function (x,y,...) or a pixel image (they must be defined for the dilated window)}

\item{meanlog}{For the distribution of radii. The logarithm of the distribution has mean equal to \code{meanlog}.}

\item{sdlog}{For the distribution of radii. The logarithm of the distribution has standard deviation equal to \code{sdlog}}

\item{seed}{Optional input (default in NULL). Is an integer passed to \code{\link{base}{set.seed}}. Used to reproduce patterns exactly.
@section Warning: A good choice of bufferdist is required and it probably depends on the distribution of radii.}
}
\value{
Returns an \code{owin} object cropped to \code{window}.
}
\description{
Simulates a Poisson point process and places discs of random radii around each point.  The radii are generated using a log normal distribution.
}
\details{
The point process needs to be simulated in a larger region than the desired window to account for the possibility of discs that intersect the window, but have germs outside the window.

The point process is generated using spatstat's \code{\link[spatstat]{rpoispp}}.
}
\examples{
w <- owin(xrange=c(0,10),yrange=c(0,10))
xi <- rboollognormdiscs(w,2,1,-1,0.5)

plot(w)
plot(xi,add=TRUE)

#For repeating experiments
w <- owin(xrange=c(0,10),yrange=c(0,10))
xi <- rboollognormdiscs(w,2,1,-1,0.5,seed=36)

plot(w)
plot(xi,add=TRUE)
}
\references{
[1] Molchanov, I. (1997) Statistics of the Boolean Model for Practitioners and Mathematicians. Wiley. (Chapter 8)

[2] **seminal paper - haven't read** Bindrich U. and Stoyan D. (1991) Stereology for pores in white bread: statistical analyses for the Boolean model by serial sections. J. Microscopy 162:231-239
}
\keyword{spatial}
