% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbpto.R
\name{rbpto}
\alias{rbpto}
\title{Simulation of Boolean Model of Grains Scaled According to a Pareto Distribution}
\usage{
rbpto(lambda, grain, win, xm, alpha, seed = NULL, xy = NULL,
  lengthscales = 1:500)
}
\arguments{
\item{lambda}{Intensity of the germ process (which is a Poisson point process)}

\item{grain}{A single owin object that gives the shape and size of the grain at scale 1}

\item{win}{The window to simulate in (an owin object)}

\item{seed}{Optional input (default in NULL). Is an integer passed to \code{\link{base}{set.seed}}. Used to reproduce patterns exactly.}

\item{xy}{A raster object that specifies pixel coordinates of the final simulated binary map.
It is used the same way as \code{xy} is \code{\link{spatstat}{as.mask}} in spatstat.
If non-null then the computations will be performed using rasters. Otherwise if \code{grain} and \code{win} are polygonal then computations may be all polygonal.}
}
\value{
An owin object.
}
\details{
The parameters xm and alpha are such that the CDF of the Pareto distribuion is \eqn{P(s <= x) = 1 - (xm / x)^{alpha}}.
Approximates the grain distribution using the scales given by \code{lengthscales} and weighted by the probability distribution function of the Pareto distribution.
}
\examples{
lambda <- 2
win <- square(r = 10)
grain <- owin(xrange = c(-0.2, 0.2), yrange = c(-0.2, 0.2))
xm <- 1
alpha <- 2
}
