% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varofobsarea.1ab.R
\name{sae.v1ab.mean}
\alias{sae.v1ab.mean}
\alias{sae.v1ab.var}
\title{Variance Estimates for Observed Area - v1ab}
\usage{
sae.v1ab.mean(xi.area, window.area, n11, n21, n12, n22)

sae.v1ab.var(xi.area, window.area, n11, n21, n12, n22)
}
\arguments{
\item{xi.area}{Observed area of the class of interest.}

\item{window.area}{Area of the observation window (do not include NA pixels)}

\item{n11}{The number of samples that were truly class 1 and fallibly classified as class 1}

\item{n21}{The number of samples that were truly class 2 and fallibly classified as class 1}

\item{n12}{The number of samples that were truly class 1 and fallibly classified as class 2}

\item{n22}{The number of samples that were truly class 2 and fallibly classified as class 2}
}
\description{
Estimates the variance of the area of a cover type observed in a thematic map created using a fallible classifier from remote sensing. 
Sampling error in the confusion matrix, assume region is a good representation of population.
}
\details{
The model for the area in the class of interest is
\deqn{
hat{|X_1|} = |\hat{X}_1|\hat{a}_{11} + |\hat{X}_2|\hat{a}_{12}
}
Where \eqn{a_{ij}} are estimates of the entry of the confusion matrix calculated by
\deqn{
\hat{a}_{ij} &:= \frac{n_{ij}}{n_{.j}}
}
where \eqn{n_{ij}} is the number of samples fallibly classed \eqn{j} and truly class \eqn{i}, and 
\eqn{n_{.j}} denotes \eqn{\sum_i^g n_{ij}}.

This model only accounts for the sampling error of the confusion matrix, and thus is kinda like assuming that the region of interest IS the population
\deqn{
\var\left[\hat{|X_1|}\right] \approx |\hat{X}_1|^2 \frac{1}{n_{.1}} \frac{1}{n_{.1} - 1} \left( n_{11} (1 - \hat{a}_{11})^2 + (n_{.1} - n_{11})(- \hat{a}_{11})^2 \right)
\quad \quad + |\hat{X}_2|^2 \frac{1}{n_{.2}} \frac{1}{n_{.2} - 1} \left( n_{12} (1 - \hat{a}_{12})^2 + (n_{.2} - n_{12})(- \hat{a}_{12})^2 \right)
}
}
\section{Functions}{
\itemize{
\item \code{sae.v1ab.var}: The variance estimate
}}

\author{
{Kassel Hingee}
}
