% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varofobsarea.1b.R
\name{sae.v1b.mean}
\alias{sae.v1b.mean}
\alias{sae.v1b.wsu.mean}
\alias{sae.v1b.var}
\alias{sae.v1b.wsu.var}
\title{Variance Estimates for Observed Area - v1b}
\usage{
sae.v1b.mean(xi.area, window.area, delta, p21, p12)

sae.v1b.wsu.mean(xi.area, window.area, delta, n11, n21, n12, n22)

sae.v1b.var(xi.area, window.area, delta, p21, p12)

sae.v1b.wsu.var(xi.area, window.area, delta, n11, n21, n12, n22)
}
\arguments{
\item{xi.area}{Observed area of the class of interest.}

\item{window.area}{Area of the observation window (do not include NA pixels)}

\item{delta}{Width of a pixel}

\item{p21}{Probability that a location fallibly classified into the cover of interest is really outside the cover of interest}

\item{p12}{Probability that a location fallibly classified into the non-interesting class is really the class of interest.}
}
\description{
Estimates the variance of the area of a cover type observed in a thematic map created using a fallible classifier from remote sensing.
}
\details{
The model for the area in the class of interest is
\deqn{|X_1| = |\hat{X}_1|-\Delta^2\text{Bin}(n_1,p_{21}) + \Delta^2\text{Bin}(n_2,p_{12})}
where $\Delta$ is the width of a pixel, $n_1$ and $n_2$ are the number of pixels in $\hat{X}_1$ and $\hat{X}_2$ respectively, and
\eqn{p_{21}} and \eqn{p_{12}} are the probability of an arbitrary point classified into class 1 or 2 being misclassified respectively (class 1 is the class of interest, class 2 is everything else).
}
\section{Functions}{
\itemize{
\item \code{sae.v1b.wsu.mean}: Expected cover area when including sampling uncertainty

\item \code{sae.v1b.var}: The variance estimate assuming confusion matrix, independent pixels error

\item \code{sae.v1b.wsu.var}: The variance estimate assuming indepedendent sampling variation in conditional confusion matrix estimate and independent pixels error
}}

\author{
{Kassel Hingee}
}
