% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varofobsarea.2d.R
\name{sae.v2d.mean}
\alias{sae.v2d.mean}
\alias{sae.v2d.wsu.mean}
\alias{sae.v2d.var}
\title{Variance Estimates for Observed Area - v2d}
\usage{
sae.v2d.mean(xi.area, window.area, delta, p21, p12, radius)

sae.v2d.wsu.mean(xi.area, window.area, delta, n11, n21, n12, n22)

sae.v2d.var(xi.area, window.area, delta, p21, p12, radius)
}
\arguments{
\item{xi.area}{Observed area of the class of interest.}

\item{window.area}{Area of the observation window (do not include NA pixels)}

\item{delta}{Width of a pixel}

\item{p21}{Probability that a location fallibly classified into the cover of interest is really outside the cover of interest}

\item{p12}{Probability that a location fallibly classified into the non-interesting class is really the class of interest.}

\item{radius}{A radius of dependence - see details for meaning.}
}
\description{
Estimates the variance of the area of a cover type observed in a thematic map created using a fallible classifier from remote sensing.
}
\details{
The model for the area in the class of interest is
\deqn{|X_1| = |\hat{X}_1|-\Delta^2$\pi r^2$\text{Bin}(n_1,p_{21}) + \Delta^2$\pi r^2$\text{Bin}(n_2,p_{12})}
where $\Delta$ is the width of a pixel, 
\eqn{p_{21}} and \eqn{p_{12}} are the probability of an arbitrary point classified into class 1 or 2 being misclassified respectively (class 1 is the class of interest, class 2 is everything else),
and 
\eqn{n_1 = \frac{|\hat{X}_1|}{\pi r^2}},
\eqn{n_2 = \frac{|\hat{X}_2|}{\pi r^2}}
where $r$ is a guess of the radius of dependence in the error process (e.g.~Pete chose $5m$) for $r$.
This is like assuming the pixels are the size of $\pi r^2$, or that error process is random on a version of the image partitioned in cells of area $\pi r^2$.
}
\section{Functions}{
\itemize{
\item \code{sae.v2d.wsu.mean}: Expected cover area when including sampling uncertainty

\item \code{sae.v2d.var}: The variance estimate assuming confusion matrix, independent error on undefined sub regions
}}

\author{
{Kassel Hingee}
}
