% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varofobsarea.3.R
\name{sae.v3.mean}
\alias{sae.v3.mean}
\alias{sae.v3.var}
\title{Variance Estimates for Observed Area}
\usage{
sae.v3.mean(xi, obswin, p21 = NA, p12 = NA)

sae.v3.var(xi, obswin, corrrad, corrstepheight, p21, p12)
}
\arguments{
\item{xi}{An observation of the RACS of interest in 1, 0, or NA valued pixels.
Pixels must be square. Must be a spatstat im object.}

\item{obswin}{Observation window}

\item{p21}{Probability that a location fallibly classified into the cover of interest is really outside the cover of interest}

\item{p12}{Probability that a location fallibly classified into the non-interesting class is really the class of interest.}

\item{corrrad}{Radius of the step function in the correlation (in the same units as xi)}

\item{corrstepheight}{Height of the step in the correlation}
}
\description{
Estimates the variance of the area of a cover type observed in a thematic map created using a fallible classifier from remote sensing.
}
\details{
To install OpenImageR had to install libtiff5-dev on my ubuntu machine
}
\section{Functions}{
\itemize{
\item \code{sae.v3.var}: The variance of the expected area using Small Area Estimation Method - Version 3. 
This methods assumes that the omission and comission errors are independent and different processes and that the correlation between errors (within each of theses processes) is a step function with radius \code{corrrad}.
}}

\examples{
xi <- as.im(heather$coarse, na.replace = 0)
obswin <- Frame(xi)
corrrad <- 5 #in units of image
corrstepheight <- 0.9
p21 <- 0.05
p12 <- 0.01
xi.sum <- sum(xi)
exparea <- sae.v3.mean(xi, obswin, p21 = p21, p12 = p12)
varguess <- sae.v3.var(xi, obswin, corrrad, corrstepheight, p21, p12)
stdofareaest <- sqrt(varguess)
}
\author{
{Kassel Hingee}
}
