% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varofobsarea.4.R
\name{sae.v4.mean}
\alias{sae.v4.mean}
\alias{sae.v4.var}
\title{Variance Estimates for Observed Area}
\usage{
sae.v4.mean(xi, obswin)

sae.v4.var(xi, obswin, erosionrad)
}
\arguments{
\item{xi}{An observation of the RACS of interest in owin form.}

\item{obswin}{Observation window in owin format}

\item{corrrad}{Radius of the step function in the correlation (in the same units as xi)}

\item{corrstepheight}{Height of the step in the correlation}

\item{p21}{Probability that a location fallibly classified into the cover of interest is really outside the cover of interest}

\item{p12}{Probability that a location fallibly classified into the non-interesting class is really the class of interest.}
}
\description{
Estimates the variance of the area of a cover type observed in a thematic map created using a fallible classifier from remote sensing.
}
\details{
To install OpenImageR had to install libtiff5-dev on my ubuntu machine
}
\section{Functions}{
\itemize{
\item \code{sae.v4.var}: A `variance' estimated by adhoc selection of erosion and dilation distances corresponding to 2*standard deviation
}}

\examples{
xi <- heather$coarse
obswin <- Frame(xi)
erosionrad <- 0.1 #in units of image
exparea <- sae.v4.mean(xi, obswin)
varguess <- sae.v4.var(xi, obswin, erosionrad)
stdofareaest <- sqrt(varguess)
}
\author{
{Kassel Hingee}
}
