% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfvlist.R
\name{summary.fvlist}
\alias{summary.fvlist}
\title{Summarise a list fv objects}
\usage{
\method{summary}{fvlist}(object, ..., na.rm = FALSE)
}
\arguments{
\item{object}{A list of fv objects}

\item{...}{Ignored.}

\item{na.rm}{If TRUE NA values in the fv object will be ignored 
(i.e. NA values will be removed from all summations and the population size will decrease by 1 for each NA value).}
}
\value{
An fv object containing the pointwise mean, pointwise sample variance, pointwise maximum and pointwise minimum.
}
\description{
This function assumes that fv objects are each realisations of the same (stochastic) object. 
It returns pointwise summaries such as observed sample mean and sample variance.
}
\examples{
obspatterns <- rpoispp(10, nsim = 10)
object <- lapply(obspatterns, Hest, W = Frame(obspatterns[[1]]))
summ <- summary.fvlist(object)

#test with NA vals
object[[1]]$km[1:46] <- NA
summ <- summary.fvlist(object, na.rm = TRUE)
}
\author{
Kassel Hingee
}
