% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlistspatialpolygonsdataframe.R
\name{unlistSpatialPolygonsDataframe}
\alias{unlistSpatialPolygonsDataframe}
\title{unlist SpatialPolygonsDataFrame}
\usage{
unlistSpatialPolygonsDataframe(spdf)
}
\arguments{
\item{spdf}{A SpatialPolygonsDataFrame}
}
\value{
a list of SpatialPolgyonsDataFrames, each data frame contains exactly one polygon.
}
\description{
Takes a SpatialPolygonsDataFrame object and splits it into individual SpatialPolygonsDataFrames. One for each polygon.
There is a nicer way: use split(, f = 1:SPDFlength).
}
