% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagPixel.R
\name{contagpixelgrid}
\alias{contagpixelgrid}
\alias{adjacency}
\title{Pixel Adjacency Contagion}
\usage{
contagpixelgrid(xi, obswin, normalise = FALSE)

adjacency(xi, obswin = NULL)
}
\arguments{
\item{xi}{An observation of a RACS of interest as a full binary map (in \code{im} format) or as the foreground set (in \code{owin} format).
In the latter case the observation window, \code{obswin}, must be supplied.
See \code{\link{stationaryracsinference-package}} for details.
  If \code{xi} is an owin object it must be of \code{mask} type.}

\item{obswin}{If \code{xi} is an \code{owin} object then \code{obswin} is an
\code{owin} object that specifies the observation window.}

\item{normalise}{If \code{TRUE} will divide result by \eqn{2 ln(2)} and add 1 to make contagion between 0 and 1 for any binary map}
}
\description{
Function for calculating the traditional contagion LPI from a binary map (O'Neill, 1988).
}
\details{
The unnormalised contagion LPI of categorical map is defined as
\deqn{\sum_i \sum_j Pij ln(Pij),} where \eqn{Pij} is the probability of
randomly selected adjacent pixels being in class \eqn{i} and class \eqn{j}
respectively, and \eqn{m} is the number of classes.

Here \eqn{m = 2} as \eqn{xi} is a binary map and we have defined 'adjacent'
pixels using the 4-neighbourhood regime.

Contagion is calculated from an adjacency matrix created using \code{adjacency}.
}
\section{Functions}{
\itemize{
\item \code{contagpixelgrid}: Traditional contagion LPI of a binary map.

\item \code{adjacency}: Calculates the adjacency matrix used in the pixel contagion
}}

\section{Warning}{
 Will fail if map is either all foreground or all background.
}

\examples{
xi <- heather$coarse
obswin <- owin(xrange = c(0,7),yrange=c(0,16))
adjmat <- adjacency(xi,obswin)
adjmat
contagion <- contagpixelgrid(xi,obswin)
contagion
#Finer resolution
xi <- heather$medium
obswin <- owin(xrange = c(0,7),yrange=c(0,16))
contagion <- contagpixelgrid(xi,obswin)
contagion
}
\references{
O'Neill, R.V., Krummel, J.R., Gardner, R.H., Sugihara, G., Jackson, B., DeAngelis, D.L., et al. (1988) Indices of landscape pattern. \emph{Landscape Ecology}, 1, 153-162.
}
