% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contagSphCont.R
\name{contagSphCont}
\alias{contagSphCont}
\title{Disc State Contagion}
\usage{
contagSphCont(XiH, XiHc, p, normalise = FALSE)
}
\arguments{
\item{XiH}{Conditional spherical contact distribution function for \eqn{\Xi}. 
Typically this is an \code{fv} object but could also be a vector of values.
In applications \code{XiH} would likely be estimated from an image using \code{\link{Hest}} in \package{spatstat}.}

\item{XiHc}{Conditional spherical contact distribution for the complement of \eqn{\Xi}. 
This is called the Conditional Core Probability in Hingee 2016.
Typically this is an \code{fv} object but could also be a vector of values.
In applications \code{XiH} would likely be estimated from an image using \code{\link{Hest}} in \package{spatstat}.}

\item{p}{The coverage fraction of \eqn{\Xi}.
In applications to images an estimate of the coverage fraction can be obtained using \code{\link{coveragefrac}}.}

\item{normalise}{Optional. If TRUE \code{contagSphCont} normalises the results so that all RACS return a value between 0 and 1. Default is FALSE.}
}
\value{
An \code{fv} object or a vector the same length as \code{XiH} corresponding to the contagion at each r value of \code{XiH}
}
\description{
Calculates the disc-state contagion LPI as described in Hingee 2016.
The disc-state contagion LPI describes the entropy (mixing) between four possible states of a disc:
\enumerate{
  \item the disc is completely contained in \eqn{\Xi}
  \item the disc does not intersect \eqn{\Xi}
  \item the centre of the disc is in \eqn{\Xi} but the disc is not contained in \eqn{Xi}
  \item the disc intersects \eqn{Xi} but the centre is outside \eqn{Xi}
}
The user is required to specify the radius of the disc which is essentially a distance for which the user is interested in quantifying mixing. 

The difference to classical contagion is that disc-state contagion is based on the spherical contact distribution instead of pixel neighbours.
One impact of this design is that a distance to quantify the mixing between \eqn{Xi} and the background is chosen by the user (for classical contagion this distance is fixed by the image resolution).
}
\details{
XiH should be a function of radius that gives (or estimates) the probability of a disc of radius \eqn{r} not intersecting \eqn{\Xi} if the disc's centre is not in \eqn{\Xi} 
\deqn{XiH(r) = P(B_r(x) \subseteq \Xi^c | x \in \Xi^c).}
Similarly \code{XiHc} should be an estimate of the probability of a disc being fully contained in \eqn{\Xi} given its centre is in \eqn{\Xi}
\deqn{XiHc(r)\approx P(B_r(x) \subseteq \Xi | x \in \Xi).}
These can both be obtained using \code{\link{Hest}} in \pkg{spatstat}.
For \code{XiHc} take care to apply Hest to the complement of \eqn{\Xi} with the observation window \eqn{W}.

If \code{XiH} and \code{XiHc} are both fv objects then they must be generated using Hest because the function automatically uses the reduce-sample border correction estimates.
In this case the return value is an fv object.

If \code{normalise} is \code{TRUE} then the result is divided by 
\eqn{\frac{-4}{e}ln(\frac{1}{e})} and added to 1 so that the normalised disc state contagion is between 0 and 1.
}
\examples{
xi <- heather$coarse
obswindow <- Frame(heather$coarse)
p <- coveragefrac(xi,Frame(xi))
XiH <- Hest(xi,W=obswindow) #Sph. Contact Distrution Estimate
XiHc <- Hest(complement.owin(xi),W=obswindow) #Conditional Core Prob. Estimate
plot(XiH,type="l",col="red") 
lines(XiHc,type="l",col="black") 

contagion <- contagSphCont(XiH,XiHc,p)
plot(contagion)

}
\references{
Hingee, K.L. (2016) Statistics for Patch Observations. ISPRS Congress Proceedings p. IPSRS.
}
