% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance.R
\name{covariance}
\alias{covariance}
\title{A spatial covariance, also known as `two-point probability', estimator for stationary RACS}
\usage{
covariance(xi, w = NULL, inclraw = FALSE, setCovBoundaryThresh = 0.1 *
  area.owin(w))
}
\arguments{
\item{xi}{An observation of the RACS of interest. It can be in \pkg{spatstat}'s \code{owin} or \code{im} format. If \code{xi} is in \code{im} format then it is assumed that the pixels will be valued 1 (for foreground), 0 (for background) and NA for unobserved.}

\item{w}{The observation window in \code{owin} format. If it isn't included and \code{xi} is an \code{owin} object then \code{w} is taken to be the smallest rectangle enclosing \code{xi}. If \code{xi} is a \code{im} object than \code{w} is all the non-NA pixels in \code{xi}.}

\item{inclraw}{If TRUE the output will be two \code{im} objects one for the standard estimator and one raw estimate.}

\item{setCovBoundaryThresh}{Any vector \eqn{v} such that set covariance of the observation window is smaller than this threshold is given a covariance of NA to avoid instabilities caused by dividing by very small areas,}
}
\value{
A \pkg{SpatStat} \code{im} object containing the estimated covariance. The grey scale values in this image represent the covariance for an array of vectors. If the raw version is requested then a list of \code{im} objects is returned.
}
\description{
These functions estimate the covariance of a stationary RACS. 
The covariance is also known as the two-point coverage probability, and very closely related to the semivariogram.
 The covariance of a vector \eqn{v} is the probability of two points separated by a \eqn{v} being covered by the random set \eqn{\Xi}
\deqn{C(v) = P(\{x,x+v\}\subseteq \Xi).}
}
\details{
The reduced sample estimator [1] is
\deqn{ \hat{C}(v) = \frac{\gamma_{W\cap X}(v)}{\gamma_W(v)}}
where \eqn{\gamma_{W}(v)} is the set covariance of the observation window \eqn{|W \cap (W\oplus v)|} and \eqn{\gamma_{W\cap X}(v)} is the set covariance of an observation, \eqn{X}, of the RACS \eqn{\Xi}, 
\eqn{\gamma_{W\cap X}(v) = |W \cap X \cap ((W\cap X ) \oplus v)|}.
The raw estimate (if requested) is
\deqn{\hat{\tilde{C}}(v) = \frac{\gamma_{W\cap X}(v)}{|W|}.}
\code{covariance} uses Fourier transforms to calculate set covariances (using the \code{\link{setcov}} function from \pkg{spatstat}). 
Vectors with small \eqn{\gamma_W(v)} are eliminated using \code{setCovBoundaryThresh} because they cause numerical instabilities.
}
\examples{
xi <- heather$coarse
covar <- covariance(xi,inclraw=FALSE)
covar <- covariance(as.im(heather$coarse,na.replace=0))
}
\references{
[1] Serra, J.P. (1982) Image Analysis and Mathematical Morphology. London; New York: Academic Press.
}
\author{
{Kassel Hingee}
}
\keyword{nonparametric}
\keyword{spatial}
