% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageprobabilityEst.R
\name{coveragefrac}
\alias{coveragefrac}
\title{Estimate the coverage fraction of a stationary RACS}
\usage{
coveragefrac(xi, w)
}
\arguments{
\item{xi}{An observation of a RACS.}

\item{w}{The window of observation (not necessarily rectangular)}
}
\value{
an estimate of the coverage probability
}
\description{
Estimates the probability of an aribitray point being inside a stationary RACS.
Because the RACS is stationary this is equivalent to estimating \eqn{P(o\in\Xi)}. 
The estimate is simply the fraction of \eqn{\Xi} in the observed window.
}
\details{
in reality the area of xi is probably slightly different - depends on the sensing method of xi. Can assume really close though.
}
\examples{
xi <- heather$coarse
obswindow <- Frame(heather$coarse)
coverageProb <- coveragefrac(xi,obswindow)
}
\author{
Kassel Hingee
}
\keyword{nonparametric}
\keyword{spatial}

