% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppicka.R
\name{cppicka}
\alias{cppicka}
\title{Picka's Modified Estimator of Coverage Probability}
\usage{
cppicka(xi, obswin = NULL, setcov_boundarythresh = NULL)
}
\arguments{
\item{xi}{An observation of a RACS in \pkg{spatstat}'s \code{owin} or \code{im} format.}

\item{obswin}{The window of observation (not necessarily rectangular) also in \code{owin} format.}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window is smaller than this threshold is given a value of NA to avoid instabilities caused by dividing by very small areas}
}
\value{
An estimate of the coverage probability
}
\description{
Picka2000va modifies coverage probability and covariance estimators to develop estimators of centered covariance and pair-correlation function that are balanced.
The entity computed by this function corresponds to Picka's \eqn{\hat{m}_1[\check{0},v]}.
}
\details{
If \code{xi} is in \code{im} format then \code{xi} must be an image of 1s, 0s and NAs
 representing inside the set, outside the set and outside the observation window respectively.
 \code{coverageprob} will not accept a \code{obswin} argument if \code{xi} is in \code{im} format.
}
\examples{
xi <- heather$coarse
obswindow <- Frame(heather$coarse)
cp <- coverageprob(xi, obswindow)
cpp1 <- cppicka(xi, obswindow)
cpp1[as.ppp(c(0,0), W = obswindow)]
plot(cpp1[cpp1 < 2, drop = FALSE])
}
\author{
Kassel Liam Hingee
}
\keyword{nonparametric}
\keyword{spatial}
