% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integratebyrad.R
\name{integratebyradius}
\alias{integratebyradius}
\title{Estimates the integral of image values within a radius r.}
\usage{
integratebyradius(centre, values)
}
\arguments{
\item{centre}{The centre location}

\item{values}{A pixel image to be integrated}
}
\value{
An fv object
}
\description{
Estimates the integral of image values within a radius r from the centre.
}
\examples{
twptprob <- racscovariance(heather$coarse,Frame(heather$coarse))
kfcn <- integratebyradius(c(0,0),twptprob)
# plot(kfcn)

#duplicate with rotmean()
twptprob.iso <- rotmean(twptprob,result="fv",Xname="radius")
kfcn3.fun <- function(R){
   integrate(as.function.fv(with.fv(twptprob.iso,f*(2*pi*.x), fun=TRUE)),lower=0,upper=R)$value
}
Rvals <- seq(0.1,8,by=0.1)
# points(Rvals,lapply(Rvals,kfcn3.fun),col="green")

}
