% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lac.R
\name{lac}
\alias{lac}
\title{Estimates of Theoretical Lacunarity}
\usage{
lac(boxes, covariance, p)
}
\arguments{
\item{boxes}{Either a list of sidelengths for square boxes or a list of \code{owin} objects of shape.}

\item{covariance}{A \code{im} object containing the covariance function (typically estimated by \code{covariance})}

\item{p}{The coverage probability. Typically estimated by the fraction of the observation window covered by the set of interest.}
}
\description{
Estimates the Theoretical Lacunarity of a stationary RACS using estimated covariance (two-point probability) and coverage fraction.
}
\details{
Denoted the estimated covariance by \eqn{\hat{C}(v)} and coverage probability \eqn{\hat{p}} then the estimate lacunarity is
\deqn{\frac{1}{\hat{p}^2 |B|^2}\int \gamma_B(v)\hat{C}(v)dv -1 }
}
\examples{
xi <- heather$coarse
covar <- covariance(xi,inclraw=FALSE)
p <- area(xi)/area(Frame(xi))
boxes <- c(0.3,0.5,0.8,1)
lac(boxes,covar,p)
otherboxes <- list(square(0.3),square(0.5),disc(0.8),square(1))
lac(otherboxes,covar,p)

#Test on a Boolean Model
lambda <- 2.2064E-3
discr <- 10
w <- owin(xrange=c(0,100),yrange=c(0,100))
xi <- rBooleanDetermDiscs(lambda,discr,w)
plot(xi)
xiimg <- as.mask(xi,eps=c(0.1,0.1))
#estimated lacunarity
estcov <- covariance(xiimg,w=w)
estp <- area(xiimg)/area(w)
lac(c(0.5,0.8,1,2,3,4,5,6),estcov,estp)
#theoretical lacunarity very different because window is small **I think
thcovariance <- thcovarDeterministicDiscs(
                 xrange=c(-10,10),
	    yrange=c(-10,10),
	    eps=c(0.1,0.1),lambda,discr)
thcoverageprob <- booldetermdiscs_truecoveragefrac(lambda,discr)
lac(c(0.5,0.8,1,2,3,4,5,6),thcovariance,thcoverageprob)

#look at why there is a difference
estp
thcoverageprob
plot(as.solist(list(estcov,thcovariance)),axes=TRUE)
}

