% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lac_gb.R
\name{lacgb}
\alias{lacgb}
\title{Gliding Box lacunarity from a black and white image}
\usage{
lacgb(img, sidelengths, inclraw = TRUE, W = Frame(img))
}
\arguments{
\item{img}{An image of 0's and 1's, and NA pixels are assumed to be outside the observation window.}

\item{sidelengths}{A list of box sidelengths in the dimensions of \code{img}}

\item{inclraw}{If TRUE the function will also return a gliding box lacunarity that ignores edge effects.}

\item{W}{Optional observation window. The observation window used for the estimator will be the union of \code{W} and the NA pixles in \code{img}.}
}
\value{
An fv object with columns for no edge correction (raw) and reduced sample border correction (RS)
}
\description{
Calculates the gliding box lacunarity
}
\details{
Calculates the gliding box lacunarity for a given range of box sizes (`radius').
The centres are given the pixels of \code{img}.

Note: (1) The sidelengths are rounded such that they are an odd number of pixels across. (2) The reduced sample points are given by Minkowski subtraction
}
\examples{
img <- as.im(heather$coarse,na.replace=0)
sidelengths <- c(0.2,1,2.2,3) #in units of img
lac <- lacgb(img,sidelengths)
plot(lac, cbind(RS,raw) ~ s)

}
\references{
The gliding box algorithm is described in Allain, C. and Cloitre, M. (1991) Characterizing the lacunarity of random and deterministic fractal sets. Physical Review A, 44, 3552-3558.
}

