% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl.R
\name{mvl}
\alias{mvl}
\title{Compute MVL estimates using all estimators}
\usage{
mvl(xiim, boxwidths, estimators = c("MVLg.mattfeldt", "MVLg.pickaint",
  "MVLg.pickaH", "MVLcc.mattfeldt", "MVLcc.pickaint", "MVLc", "MVLgb"))
}
\arguments{
\item{xiim}{A \pkg{spatstat} \code{im} object with pixel values that are either TRUE, FALSE or NA. TRUE represents foreground, FALSE respresents background and NA represents unobserved locations.}

\item{boxwidths}{A list of box boxwidths}

\item{estimators}{A list of estimator names - see details for possibilities.}
}
\value{
An \code{fv} object.
}
\description{
This function computes all the MVL estimators described in Hingee 2019** for square boxes and raster binary maps.
It calls the functions \code{MVLc}, \code{MVLg}, \code{MVLcc} and \code{MVLgb}.
}
\details{
The estimators available are
\itemize{
\item \code{"MVLc"} The unmodified (unbalanced) covariance estimator provided by \code{MVLc()}
\item \code{"MVLgb"} The Gliding-Box estimator of Alain and Cloitre **ref. Calls \code{MVLgb()}
\item \code{"MVLg.mattfeldt"} See help for \code{MVLg()}
\item \code{"MVLg.pickaint"} See help for \code{MVLg()}
\item \code{"MVLg.pickaH"} See help for \code{MVLg()}
\item \code{"MVLcc.mattfeldt"} See help for \code{MVLcc()}
\item \code{"MVLcc.pickaint"} See help for \code{MVLcc()}
\item \code{"MVLcc.pickaH"} See help for \code{MVLcc()}
}
}
\examples{
xi <- heather$coarse
xiim <- as.im(xi, value = TRUE, na.replace = FALSE)
mvlests <- mvl(xiim, seq(1, 10, by = 0.1))
}
