% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvl.R
\name{mvl}
\alias{mvl}
\alias{mvl.cvchat}
\title{Mass variance lacunarity estimatation using all estimators}
\usage{
mvl(xi, boxwidths, estimators = c("MVLg.mattfeldt", "MVLg.pickaint",
  "MVLg.pickaH", "MVLcc.mattfeldt", "MVLcc.pickaint", "MVLc", "MVLgb"),
  obswin = NULL, includenormed = FALSE,
  setcov_boundarythresh = 1e-06)

mvl.cvchat(boxwidths, estimators = c("MVLg.mattfeldt", "MVLg.pickaint",
  "MVLg.pickaH", "MVLcc.mattfeldt", "MVLcc.pickaint", "MVLc"),
  phat = NULL, cvchat = NULL, cpp1 = NULL)
}
\arguments{
\item{xi}{A binary map of an observation of a RACS of interest. See
\code{\link{stationaryracsinference-package}} for details.}

\item{boxwidths}{A list of box widths}

\item{estimators}{A list of estimator names - see details for possibilities. \code{estimators = "all"} will select all estimators.}

\item{obswin}{If \code{xi} is an \code{owin} object then \code{obswin} is an
\code{owin} object that specifies the observation window.}

\item{includenormed}{A logical value. If TRUE then MVL estimates normalised by the MVL values at zero will be included in a returned list of fv objects}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window is smaller than this threshold
is given a covariance estimate (and other similar estimate) of NA to avoid instabilities caused by dividing by very small areas.
If NULL is supplied (default) then 1E-6 is used.}

\item{phat}{Traditional estimate of coverage probability.}

\item{cvchat}{Traditional estimate of covariance (often from \code{tradcovarest}).}

\item{cpp1}{Picka's estimate of coverage probability (often from \code{cppicka}).}
}
\value{
An \code{fv} object.
}
\description{
Estimates mass variance lacunarity (MVL) using all estimators described in [Hingee2019**] from binary maps for square boxes.
It calls the functions \code{mvlc}, \code{mvlg}, \code{mvlcc} and \code{mvlgb}.
}
\details{
The function is not able to estimate MVL for non-square boxes as the gliding box estimator is included.
To estimate MVL for non-square boxes use \code{mvlcc} or \code{mvlg} directly.

If \code{xi} is in \code{owin} format then \code{obswin} and \code{xi} are converted
 into a binary map in \code{im} format using \code{\link[spatstat]{as.im}}

The estimators available are
\itemize{
\item{\code{"MVLc"}} The unmodified (unbalanced) covariance estimator provided by \code{\link{mvlc}}
\item{\code{"MVLgb"}} The Gliding-Box estimator of Allain and Cloitre [allain1991ch]. Calls \code{\link{mvlgb}}
\item{\code{"MVLg.mattfeldt"}} See help for \code{\link{mvlg}}
\item{\code{"MVLg.pickaint"}} See help for \code{\link{mvlg}}
\item{\code{"MVLg.pickaH"}} See help for \code{\link{mvlg}}
\item{\code{"MVLcc.mattfeldt"}} See help for \code{\link{mvlcc}}
\item{\code{"MVLcc.pickaint"}} See help for \code{\link{mvlcc}}
\item{\code{"MVLcc.pickaH"}} See help for \code{\link{mvlcc}}
}
}
\section{Functions}{
\itemize{
\item \code{mvl}: Computes MVL estimates from a binary map.

\item \code{mvl.cvchat}: Computes covariance-based estimator of MVL from the traditional estimate of covariance,
Picka's reduced window coverage probability estimates and the traditional coverage probability estimate.
}}

\examples{
xi <- heather$coarse
xi <- as.im(xi, value = TRUE, na.replace = FALSE)
mvlests <- mvl(xi, seq(1, 10, by = 0.1))
}
