% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlc.R
\name{mvlc}
\alias{mvlc}
\alias{mvl}
\alias{lac}
\title{Covariance-based calculations of mass variance lacunarity}
\usage{
mvlc(boxes, covariance = NULL, p = NULL, xiim = NULL)

mvl(boxes, covariance = NULL, p = NULL, xiim = NULL)

lac(boxes, covariance = NULL, p = NULL, xiim = NULL)
}
\arguments{
\item{boxes}{Either a list of sidelengths for square boxes or a list of \code{owin} objects of shape.}

\item{covariance}{A \code{im} object containing the covariance function (typically estimated by \code{covariance})}

\item{p}{The coverage probability. Typically estimated by the fraction of the observation window covered by the set of interest.}

\item{xiim}{An observation of a stationary RACS in \code{im} format. \code{xiim} must have values of either 1, 0 or NA; 1 denotes inside the RACS, 0 denotes outside, and NA denotes unobserved.}
}
\value{
Either an \code{fv} object containing the MVL values and box side lengths or if \code{boxes} is a list of owin objects then \code{lac} returns a list of corresponding MVL values. Note if NA or NaN values in the \code{covariance} object are used then function will return NA or NaN instead of an MVL value.
}
\description{
Estimates the mass variance lacunarity (MVL) of a stationary RACS from a bi-tonal image, or calculates the MVL from provided covariance (two-point probability) and coverage probability.
}
\details{
If we denoted the estimated covariance by \eqn{\hat{C}(v)} and coverage probability \eqn{\hat{p}} then the estimate of MVL is
\deqn{\frac{1}{\hat{p}^2 |B|^2}\int \gamma_B(v)\hat{C}(v)dv -1 }
}
\examples{
xi <- heather$coarse
covar <- covariance(xi,inclraw=FALSE)
p <- area(xi)/area(Frame(xi))
sidelengths <- seq(0.3,14,by=0.2)
plot(lac(sidelengths,covar,p))
# what is the MVL estimates for boxes that are discs?
discboxes <- lapply(sidelengths/2,disc)
discmvls <- mvlc(discboxes,covar,p)
points(sidelengths,discmvls)
}
