% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlcc.R
\name{mvlcc}
\alias{mvlcc}
\title{Centred covariance based estimates of mass variance lacunarity}
\usage{
mvlcc(boxes, cencovar = NULL, p = NULL, xiim = NULL,
  estimator = "pickaH")
}
\arguments{
\item{boxes}{Either a list of sidelengths for square boxes or a list of \code{owin} objects of any shape.}

\item{cencovar}{A \code{im} object containing the centred covariance function}

\item{p}{The coverage probability. Typically estimated by the fraction of the observation window covered by the set of interest.}

\item{xiim}{An observation of a stationary RACS in \code{im} format. \code{xiim} must have values of either 1, 0 or NA; 1 denotes inside the RACS, 0 denotes outside, and NA denotes unobserved.}

\item{estimator}{If an observation \code{xiim} is passed then \code{estimator} will select the balancing method that \code{ccvc} uses to estimate the centred covariance.}
}
\value{
If \code{boxes} is a list of numerical values then MVL is estimated for square boxes with side length given by \code{boxes}.
 The returned object is then an \code{fv} object containing estimates of MVL, box mass variance and box mass mean.
 
 If \code{boxes} is a list of owin objects then \code{mvlcc} returns a dataframe of with columns corresponding to estimates of MVL, box mass variance and box mass mean.
 Note if NA or NaN values in the \code{covariance} object are used then \code{mvlc} will return NA or NaN instead of an MVL value.
}
\description{
Estimates the mass variance lacunarity (MVL) of a stationary RACS using centred covariance estimates.
 The centred covariance and coverage probability can be provided or estimated from binary map.
}
\details{
If we denoted the estimated centred covariance by
\eqn{\hat{\kappa}(v)}{k(v)} and coverage probability \eqn{\hat{p}}{p} then
the estimate of MVL is
\deqn{ \frac{1}{\hat{p}^2 |B|^2}\int \gamma_B(v)\hat{\kappa}(v)dv }{ \int gammaB(v) k(v) dv  /  (p^2 |B|^2).}
}
\examples{
xi <- heather$coarse
cencovar <- cencovariance(xi, obswin = Frame(xi), estimators = c("pickaH"), drop = TRUE)
p <- area(xi) / area(Frame(xi))
sidelengths <- seq(0.3, 14, by = 0.2)
mvlccest <- mvlcc(sidelengths, cencovar, p)
# what is the MVL estimates for boxes that are discs?
discboxes <- lapply(sidelengths / 2, disc)
discmvls <- mvlcc(discboxes, cencovar, p)
# points(sidelengths, discmvls)

#direct to an image
xiim <- as.im(xi, na.replace = 0)
mvlccest <- mvlcc(sidelengths, xiim = xiim, estimator = "pickaH")

}
\keyword{nonparametric}
\keyword{spatial}
