% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlcc.R
\name{mvlcc}
\alias{mvlcc}
\title{Centred covariance based calculations of mass variance lacunarity}
\usage{
mvlcc(boxes, cencovar = NULL, p = NULL, xiim = NULL,
  modification = "pickaH")
}
\arguments{
\item{boxes}{Either a list of sidelengths for square boxes or a list of \code{owin} objects of any shape.}

\item{cencovar}{A \code{im} object containing the centred covariance function}

\item{p}{The coverage probability. Typically estimated by the fraction of the observation window covered by the set of interest.}

\item{xiim}{An observation of a stationary RACS in \code{im} format. \code{xiim} must have values of either 1, 0 or NA; 1 denotes inside the RACS, 0 denotes outside, and NA denotes unobserved.}

\item{modification}{If an observation \code{xiim} is passed then \code{modification} will select the balancing method that \code{ccvc} uses to estimate the centred covariance.}
}
\value{
Either an \code{fv} object containing the MVL values and box side lengths or if \code{boxes} is a list of owin objects then \code{mvlc} returns a list of corresponding MVL values.
 Note if NA or NaN values in the \code{cencovar} object are used then \code{mvlc} will return NA or NaN instead of an MVL value.
}
\description{
Estimates the mass variance lacunarity (MVL) of a stationary RACS from a bi-tonal image using centred covariance estimates.
 The centred covariance and coverage probability can be provided or estimated from an image.
}
\details{
\code{mvl} and \code{lac} are aliases of mvlc.
If we denoted the estimated centred covariance by \eqn{\hat{\kappa}(v)} and coverage probability \eqn{\hat{p}} then the estimate of MVL is
\deqn{\frac{1}{\hat{p}^2 |B|^2}\int \gamma_B(v)\hat{kapp}(v)dv}
}
\examples{
xi <- heather$coarse
cencovar <- ccvcs(xi, obswin = Frame(xi), modifications = c("pickaH"))$pickaH
p <- area(xi) / area(Frame(xi))
sidelengths <- seq(0.3, 14, by = 0.2)
# plot(mvlcc(sidelengths, cencovar, p))
# what is the MVL estimates for boxes that are discs?
discboxes <- lapply(sidelengths / 2, disc)
discmvls <- mvlcc(discboxes, cencovar, p)
# points(sidelengths, discmvls)

#direct to an image
xiim <- as.im(xi, na.replace = 0)
# plot(mvlcc(sidelengths, xiim = xiim, modification = "pickaH"))
# plot(add = TRUE, mvlc(sidelengths, xiim = xiim), lty = "dashed", col = "red")

}
\keyword{nonparametric}
\keyword{spatial}
