% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlg.R
\name{mvlg}
\alias{mvlg}
\title{Pair-correlation based estimates of mass variance lacunarity}
\usage{
mvlg(boxes, paircorr = NULL, xiim = NULL)
}
\arguments{
\item{boxes}{Either a list of sidelengths for square boxes or a list of \code{owin} objects of any shape.}

\item{paircorr}{A \code{im} object containing the pair-correlation function}

\item{xiim}{An observation of a stationary RACS in \code{im} format. \code{xiim} must have values of either 1, 0 or NA; 1 denotes inside the RACS, 0 denotes outside, and NA denotes unobserved.}
}
\value{
If \code{boxes} is a list of numerical values then MVL is estimated for square boxes with side length given by \code{boxes}.
 The returned object is then an \code{fv} object containing estimates of MVL.
 If \code{boxes} is a list of owin objects then \code{mvlg} returns a dataframe of with columns corresponding to estimates of MVL.
 
 Note if NA or NaN values in the \code{paircorr} object are used then \code{mvlg} will return NA or NaN instead of an MVL value.
}
\description{
Estimates the mass variance lacunarity (MVL) of a stationary RACS by estimating pair-correlation from a binary map.
 It can also calculate the MVL of a RACS from a provided pair-correlation function.
}
\details{
If we denoted the estimated pair-correlation by \eqn{\hat{g}(v)}{g(v)} then the estimate of MVL is
\deqn{\frac{1}{|B|^2}\int \gamma_B(v)\hat{g}(v)dv - 1. }{  \int gammaB(v) g(v) dv  /  (|B|^2)  - 1.  }
}
\examples{
xi <- heather$coarse
pcln <- paircorr(as.im(xi, na.replace = 0), estimators = "pickaH", drop = TRUE)
sidelengths <- seq(0.3, 14, by = 0.2)
mvlgest <- mvlg(sidelengths, pcln)
# what is the MVL estimates for boxes that are discs?
discboxes <- lapply(sidelengths / 2, disc)
discmvls <- mvlg(discboxes, pcln)
# points(sidelengths, discmvls)

}
\keyword{nonparametric}
\keyword{spatial}
