% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgb.R
\name{mvlgb}
\alias{mvlgb}
\title{Gliding Box estimation of mass variance lacunarity}
\usage{
mvlgb(boxwidths, xiim, obswin = Frame(xiim))
}
\arguments{
\item{boxwidths}{A list of suggested box widths in the same units as \code{xiim}. 
Note the actual box widths used by \code{mvlgb} will be the closest multiple of an odd number of pixel widths.}

\item{xiim}{An image of pixels valued either \code{0}, \code{1} or \code{NA}. \code{NA} valued pixels are assumed to be outside the observation window.}

\item{obswin}{Optional observation window. The observation window used for the estimator will be the intersection of \code{obswin} and the pixels that are not \code{NA} in \code{xiim}.}
}
\value{
An \code{fv} object containing estimates of MVL, box mass variance and box mass mean computed using the gliding box estimator described in [1]. 
 The box widths (labelled \code{s}) are always odd multiples of the pixel width.
}
\description{
Calculates the gliding box estimate [1] of mass variance lacunarity from a binary map.
}
\details{
Calculates the gliding box estimate [1] of mass variance lacunarity for a given range of square box sizes. 
The algorithm uses the pixel locations in \code{xiim} as an array of box centre locations to compute
 the mean and variance of the area in a random box of a given size.
Locations where the box is not completely within the observation window are ignored.
}
\section{WARNING}{
 \code{mvlgb} uses the \code{\link[RcppRoll]{roll_sum}} function in \pkg{RcppRoll} to operate.
\pkg{RcppRoll} must be installed.

Note: The side lengths are rounded such that they are an odd number of pixels across.
}

\examples{
xiim <- as.im(heather$coarse, na.replace = 0)
boxwidths <- seq(0.2, 14, by = 0.2) #in units of xiim
mvlest <- mvlgb(boxwidths, xiim)

}
\references{
[1] Allain, C. and Cloitre, M. (1991) Characterizing the lacunarity of random and deterministic fractal sets. Physical Review A, 44, 3552-3558.
}
\keyword{nonparametric}
\keyword{spatial}
