% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvlgb.R
\name{mvlgb}
\alias{lacgb}
\alias{mvlgb}
\title{Gliding Box lacunarity from a black and white image}
\usage{
mvlgb(img, sidelengths, inclraw = FALSE, W = Frame(img), method = "")

lacgb(img, sidelengths, inclraw = FALSE, W = Frame(img), method = "")
}
\arguments{
\item{img}{An image of 0's and 1's, and NA pixels are assumed to be outside the observation window.}

\item{sidelengths}{A list of suggested box side lengths in the dimensions of \code{img}. Note the actual side lengths used will be the closest multiple of an odd number of pixel widths.}

\item{inclraw}{If TRUE the function will also return a gliding box lacunarity that ignores edge effects. (Default is FALSE)}

\item{W}{Optional observation window. The observation window used for the estimator will be the union of \code{W} and the NA pixles in \code{img}.}

\item{method}{Obsolete.}
}
\value{
An fv object with columns for no edge correction (raw) and reduced sample border correction (RS). The side lengths (labelled \code{s}) are always odd multiples of pixel widths.
}
\description{
Calculates the gliding box lacunarity
}
\details{
Calculates the gliding box lacunarity for a given range of box sizes (`radius').
The centres are given the pixels of \code{img}. If the package \code{raster} is available a moving window algorithm is may be used (via \code{focal()}), however for now the convolution method in spatstat appears to be faster and is default.

Note: (1) The sidelengths are rounded such that they are an odd number of pixels across. (2) The reduced sample points are given by Minkowski subtraction
}
\examples{
img <- as.im(heather$coarse,na.replace=0)
sidelengths <- c(0.2,1,2.2,10) #in units of img
lac <- mvlgb(img,sidelengths, inclraw=TRUE)
plot(lac, cbind(RS,raw) ~ s)

}
\references{
The gliding box algorithm is described in Allain, C. and Cloitre, M. (1991) Characterizing the lacunarity of random and deterministic fractal sets. Physical Review A, 44, 3552-3558.
}

