% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcln.R
\name{paircorr}
\alias{paircorr}
\alias{paircorr.cvchat}
\title{Balanced estimation of pair-correlation.}
\usage{
paircorr(xi, obswin = NULL, setcov_boundarythresh = NULL,
  estimators = "all", drop = FALSE)

paircorr.cvchat(cvchat, cpp1 = NULL, phat = NULL, estimators = "all",
  drop = FALSE)
}
\arguments{
\item{xi}{A binary map of an observation of a RACS of interest. See
\code{\link{stationaryracsinference-package}} for details.}

\item{obswin}{If \code{xi} is an \code{owin} object then \code{obswin} is an
\code{owin} object that specifies the observation window.}

\item{setcov_boundarythresh}{Any vector \eqn{v} such that set covariance of the observation window
is smaller than this threshold is given a covariance of NA to avoid instabilities caused by dividing by very small areas,}

\item{estimators}{A list of strings specifying estimators to use. 
See details.
\code{estimators = "all"} will select all available estimators.}

\item{drop}{If TRUE and one estimator selected then the returned value will be a single \code{im} object and not a list of \code{im} object.
estimators = "all" will select all inbuilt estimators. See details.}

\item{cvchat}{The traditional estimate of covariance in \code{im} format. 
Typically created with \code{\link{tradcovarest}}.}

\item{cpp1}{Picka's reduced window estimate of coverage probability in \code{im} format - used in improved (balanced) covariance estimators.
Can be generated using \code{\link{cppicka}}.}

\item{phat}{The traditional estimate of coverage probability,
which is the observed foreground area in \code{xi} divided by the total area of the observation window.
See \code{\link{coverageprob}} for more information.}
}
\value{
If \code{drop = TRUE} and a single estimator requested then a
  \code{im} object containing the pair-correlaion estimate. Otherwise a
  named \code{imlist} of \code{im} objects containing the pair-correlation
  estimates for each requested estimator.
}
\description{
Estimates the pair-correlation function of a stationary RACS. 
The traditional pair-correlation estimator and three 'balanced' estimators suggested by Picka
are available.
}
\details{
The pair-correlation of a stationary RACS is 
\deqn{g(v) = C(v) / p^2.}

  The estimators available are (see [Chapter 4, hingee2019thesis] for
  more information): 
  \itemize{ 
  \item{\code{trad}} the traditional pair-correlation estimator which is \eqn{Chat(v) / (phat^2)}, where \eqn{Chat} and \eqn{phat} are 
the traditional estimates of coverage probability and covariance respectively. 
  \item{\code{mattfeldt}} an `intrinsically' balanced pair-correlation estimator suggested by Picka.
  A similar isotropic pair-correlation estimator was later studied by Mattfeldt and Stoyan [mattfeldt2000im].
  \item{\code{pickaint}} Picka's 'intrinsically' balanced pair-correlation estimator [picka2000va]. 
  \item{\code{pickaH}} Picka's 'additively' balanced pair-correlation estimator [picka2000va].
  }
}
\section{Functions}{
\itemize{
\item \code{paircorr}: Estimates pair-correlation from a binary map.

\item \code{paircorr.cvchat}: Generates pair-correlation estimates from
a traditional estimate of covariance, Picka's reduced window estimate of coverage probability,
and the traditional estimate of coverage probability.
If these estimates already exist then \code{paircorr.cvchat} can save significant computation time.
}}

\examples{
xi <- heather$coarse
#estimate directly from a binary map
pclns_direst <- paircorr(as.im(xi, na.replace = 0), estimators = "all")

#estimate using traditional covariance estimates, traditional coverage
#probability estimate and Picka's reduced window coverage probability
#estimates.
obswin <- Frame(xi)
phat <- coverageprob(xi, obswin = Frame(xi))
cvchat <- tradcovarest(xi)
cpp1 <- cppicka(xi, obswin = Frame(heather$coarse))
pclns_frcvc <- paircorr.cvchat(cvchat, cpp1, phat, estimators = "all")
}
\author{
{Anonymous}
}
\keyword{nonparametric}
\keyword{spatial}
