\name{An overview of this package}
\alias{racssummfuncs-package}
\alias{racssummfuncs}
\docType{package}
\title{
\packageTitle{racssummfuncs}: An Overview
}
\description{
\packageDescription{racssummfuncs}
}
\details{
This package contains functions that estimate various parameters/properties of random closed sets (RACS) from single 2D bi-tonal digital images.
 It enables description of spatial patterns from a rigorous RACS foundation.
 The theoretical background and motivation for these tools can be found in \cite{Hingee (2016)}, \cite{Chiu et. al. (2013)} and forthcoming papers.
 An introduction to using these tools on remotely sensed data is included in  the vignette \emph{remotesensedatademo}. 
 This can be accessed by starting \R and typing \kbd{library(racssummfuncs)} and \kbd{vignette("remotesensedatademo")}.
 Specific help for each function can be accessed by \kbd{help(fun)} where \kbd{fun} is replaced by the name of the function.

The functions accept observations in raster \code{owin} format provided by the \pkg{spatstat} package, or for \command{mvlc} and \code{mvlgb} the observation can be provided as an \code{im} object (also provided by \pkg{spatstat}) with pixel values of only \eqn{1}s, \eqn{0}s and \code{NA}s.
 \pkg{spatstat} has many tools that ease the statistical analysis of spatial data, and many of the functions in \pkg{racssummfuncs} take advantage of these. 
 Conversion from other raster formats to \code{im} is quite easy using either \code{\link[spatstat]{im}} or \code{\link[spatstat]{as.im}}.
 The function \code{\link[spatstat]{as.owin}} can then be used to obtain an \code{owin} object; note that for the \code{owin} object to correspond to all the \eqn{1}s in the \code{im} object, every non-\eqn{1} pixel must be first set to \code{NA}. 
 
Typically the functions will also require passing an \code{owin} object that specifies the observation window.

The summary functions that this package can estimate include:
\enumerate{
\item coverage probability (coverage fraction)
\item covariance (two-point probability function)
\item mass variance lacunarity
\item spherical contact contagion.
}
Another important summary function, spherical contact distribution (used heavily in \cite{Hingee (2016)}), can be estimated using \pkg{spatstat}'s \code{\link[spatstat]{Hest}} function.
 The estimated spherical contact distribution is closely related to the core area pattern metric and the Bouligand-Minkowski method for estimating the box-counting fractal dimension.

The methods assume that the observation, \eqn{\xi}, in window \eqn{W}, is generated by a stationary mixing RACS, \eqn{\Xi}, and that \eqn{W} is sufficiently large compared to the mixing scale of \eqn{\Xi} so that \eqn{xi} is like multiple independent realisations of \eqn{\Xi}.

Some functions to assist simulation of RACS, namely Boolean models (\code{\link{rbdd}}, \code{\link{rboollognormdiscs}}) and germ grain models (\code{\link{rgermgrain}}), are also included.
A Boolean model is a process that creates a pattern by placing independent identically distributed objects (\emph{grains}) at completely spatially random locations. 
It is a two stage model, first the locations of grains, the germs are randomly distributed according to a Poisson point process, then a random grain is placed on each germ independently.
 A thorough introduction can be found in \cite{Chiu et. al. (2013)}.
}
\section{Notation}{
Through out this manual an observation (a bi-tonal \code{im} object, or \code{owin} object with an observation region) is referred to as \sQuote{\eqn{\xi}} or \sQuote{xi}.
 The underlying RACS from which this observation comes is referred to as \sQuote{\eqn{\Xi}} or \sQuote{Xi}.
}

\author{
\packageAuthor{racssummfuncs}

Maintainer: \packageMaintainer{racssummfuncs}
}
\references{
Chiu, S.N., Stoyan, D., Kendall, W.S. and Mecke, J. (2013) Stochastic Geometry and Its Applications, 3rd ed. Chichester, United Kingdom: John Wiley & Sons.

Hingee, K.L. (2016) Statistics for Patch Observations. ISPRS - International Archives of the Photogrammetry, Remote Sensing and Spatial Information Sciences pp. 235-242. ISPRS.
}
\keyword{ package }
\keyword{ spatial }
\concept{ landscape ecology }
\concept{ pattern metric }
\concept{ landscape pattern indices }
\examples{
xi <- heather$coarse    #an observation of Xi
obswindow <- Frame(heather$coarse)
 
#estimate coverage probability
coverageprob <- coverageprob(xi,obswindow)

#estimate covariance from xi
xi.covariance <- covariance(xi)
#estimate spherical contact distribution, core probability and disc-state contagion
scdest <- Hest(xi,W=obswindow)   #Hest is from spatstat package
coreprob <- Hest(complement.owin(xi),W=obswindow)
discstatecontag <- contagSphCont(scdest,coreprob,coverageprob)

plot(scdest, main="Conditional Spherical Contact Distribution Estimates\n
 Showing Estimates Using Various Border Corrections")
plot(coreprob, main = "Conditional Core Probability Estimates\n
 Showing Estimates using various Border Corrections")
plot(discstatecontag, main = "A Disc-State Contagion Estimate\n
 (Using the Reduced-Sample Border Correction)")

#Mass-Variance Lacunarity Estimates.
#If RcppRoll is installed than the gliding box estimate can also be generated for comparison.
xi.mvlc <- mvlc(c(1,2,3,4), covariance = xi.covariance, p = coverageprob)
xi.im <- as.im(xi, na.replace = 0)
xi.mvlgb <- mvlgb(xi.im, c(1,2,3,4))

#Simulation of Boolean Models
booldisc.rfixed.xi <- rbdd(2.2064E-1,0.5,obswindow)
plot(obswindow, main="A Simulation of a Boolean Model")
plot(add=TRUE, booldisc.rfixed.xi)
}
