% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbpto.R
\name{rbpto}
\alias{rbpto}
\alias{bpto.coverageprob}
\alias{bpto.germintensity}
\alias{bpto.covar}
\title{Simulation of Boolean Model of Grains Scaled According to a Pareto Distribution}
\usage{
rbpto(lambda, grain, win, xm, alpha, seed = NULL, xy = NULL,
  lengthscales = 1:500)

bpto.coverageprob(lambda, grain, xm, alpha, lengthscales = 1:500)

bpto.germintensity(coverp, grain, xm, alpha, lengthscales = 1:500)

bpto.covar(lambda, grain, xm, alpha, lengthscales = 1:500, xy)
}
\arguments{
\item{lambda}{Intensity of the germ process (which is a Poisson point process)}

\item{grain}{A single owin object that gives the shape and size of the grain at scale 1}

\item{win}{The window to simulate in (an owin object)}

\item{xm}{A parameter governing the shape of the Pareto distribution used - see details}

\item{alpha}{A parameter governing the shape of the Pareto distribution used - see details}

\item{seed}{Optional input (default in NULL). Is an integer passed to \code{\link{base}{set.seed}}. Used to reproduce patterns exactly.}

\item{xy}{A raster object that specifies pixel coordinates of the final simulated binary map.
It is used the same way as \code{xy} is \code{\link{spatstat}{as.mask}} in spatstat.
If non-null then the computations will be performed using rasters. Otherwise if \code{grain} and \code{win} are polygonal then computations may be all polygonal.}

\item{lengthscales}{A list of scales of the \code{grain} for which to approximate the Pareto distribution: The grain for a germ is chosen by selecting a scaled version of \code{grain} where \code{lengthscales} specifies the possible scales and the Pareto distribution is used to specify the probability of selection of each scale.}
}
\value{
An owin object.
}
\description{
Functions for simulation and computing theoretical values of a Boolean model with identically shaped grains with size given by an approximate (in some cases a terrible approximation) Pareto distribution.
}
\details{
The parameters xm and alpha are such that the CDF of the Pareto distribuion is \eqn{P(s <= x) = 1 - (xm / x)^{alpha}}.
Approximates the grain distribution using the scales given by \code{lengthscales} and weighted by the probability distribution function of the Pareto distribution.
}
\section{Functions}{
\itemize{
\item \code{bpto.coverageprob}: The coverage probability of the Boolean model with scaled grains distributed according to Pareto distribution. Uses approximation of truncated length scales.

\item \code{bpto.germintensity}: Computes the germ intensity of the Boolean model with scaled grains distributed according to Pareto distribution and given coverage probability. Uses approximation of truncated length scales.

\item \code{bpto.covar}: The racs covariance of the Boolean model with scaled grains distributed according to Pareto distribution. 
Uses approximation of truncated length scales.
xy is required to specify resolution and offset of pixel grid
}}

\examples{
lambda <- 1
win <- square(r = 10)
#grain <- owin(xrange = c(-0.2, 0.2), yrange = c(-0.2, 0.2))
grain <- disc(r = 0.2)
xm <- 0.01
alpha <- 2
lengthscales <- seq(1, 10, by = 0.1)

xi <- rbpto(lambda, grain, win, xm, alpha, lengthscales = lengthscales)
# plot(xi, col = "black")
# plot(win, add = TRUE)

bpto.coverageprob(lambda, grain, xm, alpha, lengthscales = lengthscales)
covar <- bpto.covar(lambda, grain, xm, alpha, lengthscales = lengthscales,
   xy = as.mask(win, eps = 0.1))
}
