% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfvlist.R
\name{summary.fvlist}
\alias{summary.fvlist}
\title{Summarise a list fv objects}
\usage{
\method{summary}{fvlist}(object, ...)
}
\arguments{
\item{object}{A list of fv objects}

\item{...}{Ignored.}
}
\value{
An fv object containing the pointwise mean, variance and maxima and minima.
}
\description{
This function assumes that fv objects are each realisations of the same (stochastic) object. 
It returns pointwise summaries such as observed sample mean and sample variance.
}
\examples{
obspatterns <- rpoispp(10, nsim = 10)
object <- lapply(obspatterns, Hest, W = Frame(obspatterns[[1]]))
summ <- summary.fvlist(object)
}
\author{
Kassel Hingee
}
